Query total cumulative cases with ability to specify location and date range of returned data.

Query total cumulative cases with ability to specify location and date range of returned data.

get_cases(
  loc = "prov",
  date = NULL,
  after = "2020-01-01",
  before = Sys.Date(),
  datetime_type = TRUE
)

Arguments

loc

string: Specify geographic filter and aggregation of returned data. Valid loc arguments are: 'canada', 'prov' and two-letter province codes (e.g. 'ON', 'BC', etc.)

date

string: If not None, return data from the specified date YYYY-MM-DD. Superceeds 'after' and 'before' parameters.

after

string: Return data on and after the specified date YYYY-MM-DD.

before

string: Return data on and before the specified date YYYY-MM-DD.

datetime_type

boolean: Return date column as a character (FALSE) or as a Date (TRUE).

Value

Dataframe containing content of API response.

Examples

get_cases(loc = "ON", before = "2021-12-31")
#> No encoding supplied: defaulting to UTF-8.
#>     cases cumulative_cases       date province
#> 1       1                1 2020-01-25  Ontario
#> 2       0                1 2020-01-26  Ontario
#> 3       1                2 2020-01-27  Ontario
#> 4       0                2 2020-01-28  Ontario
#> 5       0                2 2020-01-29  Ontario
#> 6       0                2 2020-01-30  Ontario
#> 7       1                3 2020-01-31  Ontario
#> 8       0                3 2020-02-01  Ontario
#> 9       0                3 2020-02-02  Ontario
#> 10      0                3 2020-02-03  Ontario
#> 11      0                3 2020-02-04  Ontario
#> 12      0                3 2020-02-05  Ontario
#> 13      0                3 2020-02-06  Ontario
#> 14      0                3 2020-02-07  Ontario
#> 15      0                3 2020-02-08  Ontario
#> 16      0                3 2020-02-09  Ontario
#> 17      0                3 2020-02-10  Ontario
#> 18      0                3 2020-02-11  Ontario
#> 19      0                3 2020-02-12  Ontario
#> 20      0                3 2020-02-13  Ontario
#> 21      0                3 2020-02-14  Ontario
#> 22      0                3 2020-02-15  Ontario
#> 23      0                3 2020-02-16  Ontario
#> 24      0                3 2020-02-17  Ontario
#> 25      0                3 2020-02-18  Ontario
#> 26      0                3 2020-02-19  Ontario
#> 27      0                3 2020-02-20  Ontario
#> 28      0                3 2020-02-21  Ontario
#> 29      0                3 2020-02-22  Ontario
#> 30      1                4 2020-02-23  Ontario
#> 31      0                4 2020-02-24  Ontario
#> 32      0                4 2020-02-25  Ontario
#> 33      1                5 2020-02-26  Ontario
#> 34      1                6 2020-02-27  Ontario
#> 35      2                8 2020-02-28  Ontario
#> 36      3               11 2020-02-29  Ontario
#> 37      4               15 2020-03-01  Ontario
#> 38      0               15 2020-03-02  Ontario
#> 39      5               20 2020-03-03  Ontario
#> 40      0               20 2020-03-04  Ontario
#> 41      2               22 2020-03-05  Ontario
#> 42      6               28 2020-03-06  Ontario
#> 43      0               28 2020-03-07  Ontario
#> 44      4               32 2020-03-08  Ontario
#> 45      3               35 2020-03-09  Ontario
#> 46      2               37 2020-03-10  Ontario
#> 47      5               42 2020-03-11  Ontario
#> 48     17               59 2020-03-12  Ontario
#> 49     20               79 2020-03-13  Ontario
#> 50     24              103 2020-03-14  Ontario
#> 51     42              145 2020-03-15  Ontario
#> 52     32              177 2020-03-16  Ontario
#> 53     12              189 2020-03-17  Ontario
#> 54     25              214 2020-03-18  Ontario
#> 55     44              258 2020-03-19  Ontario
#> 56     60              318 2020-03-20  Ontario
#> 57     59              377 2020-03-21  Ontario
#> 58     48              425 2020-03-22  Ontario
#> 59     78              503 2020-03-23  Ontario
#> 60     85              588 2020-03-24  Ontario
#> 61     99              687 2020-03-25  Ontario
#> 62    170              857 2020-03-26  Ontario
#> 63    135              992 2020-03-27  Ontario
#> 64    151             1143 2020-03-28  Ontario
#> 65    211             1354 2020-03-29  Ontario
#> 66    350             1704 2020-03-30  Ontario
#> 67    260             1964 2020-03-31  Ontario
#> 68    867             2831 2020-04-01  Ontario
#> 69    438             3269 2020-04-02  Ontario
#> 70    395             3664 2020-04-03  Ontario
#> 71    339             4003 2020-04-04  Ontario
#> 72    426             4429 2020-04-05  Ontario
#> 73    381             4810 2020-04-06  Ontario
#> 74    525             5335 2020-04-07  Ontario
#> 75    473             5808 2020-04-08  Ontario
#> 76    557             6365 2020-04-09  Ontario
#> 77    417             6782 2020-04-10  Ontario
#> 78    454             7236 2020-04-11  Ontario
#> 79    462             7698 2020-04-12  Ontario
#> 80    510             8208 2020-04-13  Ontario
#> 81    551             8759 2020-04-14  Ontario
#> 82    522             9281 2020-04-15  Ontario
#> 83    531             9812 2020-04-16  Ontario
#> 84    626            10438 2020-04-17  Ontario
#> 85    570            11008 2020-04-18  Ontario
#> 86    545            11553 2020-04-19  Ontario
#> 87    648            12201 2020-04-20  Ontario
#> 88    497            12698 2020-04-21  Ontario
#> 89    703            13401 2020-04-22  Ontario
#> 90    581            13982 2020-04-23  Ontario
#> 91    491            14473 2020-04-24  Ontario
#> 92    421            14894 2020-04-25  Ontario
#> 93    484            15378 2020-04-26  Ontario
#> 94    426            15804 2020-04-27  Ontario
#> 95    486            16290 2020-04-28  Ontario
#> 96    471            16761 2020-04-29  Ontario
#> 97    398            17159 2020-04-30  Ontario
#> 98    547            17706 2020-05-01  Ontario
#> 99    456            18162 2020-05-02  Ontario
#> 100   242            18404 2020-05-03  Ontario
#> 101   550            18954 2020-05-04  Ontario
#> 102   383            19337 2020-05-05  Ontario
#> 103   440            19777 2020-05-06  Ontario
#> 104   489            20266 2020-05-07  Ontario
#> 105   399            20665 2020-05-08  Ontario
#> 106   340            21005 2020-05-09  Ontario
#> 107   318            21323 2020-05-10  Ontario
#> 108   357            21680 2020-05-11  Ontario
#> 109   346            22026 2020-05-12  Ontario
#> 110   352            22378 2020-05-13  Ontario
#> 111   349            22727 2020-05-14  Ontario
#> 112   393            23120 2020-05-15  Ontario
#> 113   372            23492 2020-05-16  Ontario
#> 114   328            23820 2020-05-17  Ontario
#> 115   322            24142 2020-05-18  Ontario
#> 116   503            24645 2020-05-19  Ontario
#> 117   439            25084 2020-05-20  Ontario
#> 118   405            25489 2020-05-21  Ontario
#> 119   495            25984 2020-05-22  Ontario
#> 120   474            26458 2020-05-23  Ontario
#> 121   339            26797 2020-05-24  Ontario
#> 122   419            27216 2020-05-25  Ontario
#> 123   318            27534 2020-05-26  Ontario
#> 124   317            27851 2020-05-27  Ontario
#> 125   385            28236 2020-05-28  Ontario
#> 126   370            28606 2020-05-29  Ontario
#> 127   325            28931 2020-05-30  Ontario
#> 128   372            29303 2020-05-31  Ontario
#> 129   384            29687 2020-06-01  Ontario
#> 130   472            30159 2020-06-02  Ontario
#> 131   369            30528 2020-06-03  Ontario
#> 132   336            30864 2020-06-04  Ontario
#> 133   421            31285 2020-06-05  Ontario
#> 134   250            31535 2020-06-06  Ontario
#> 135   474            32009 2020-06-07  Ontario
#> 136   316            32325 2020-06-08  Ontario
#> 137   278            32603 2020-06-09  Ontario
#> 138   258            32861 2020-06-10  Ontario
#> 139   239            33100 2020-06-11  Ontario
#> 140   206            33306 2020-06-12  Ontario
#> 141   242            33548 2020-06-13  Ontario
#> 142   179            33727 2020-06-14  Ontario
#> 143   206            33933 2020-06-15  Ontario
#> 144   194            34127 2020-06-16  Ontario
#> 145   190            34317 2020-06-17  Ontario
#> 146   189            34506 2020-06-18  Ontario
#> 147   205            34711 2020-06-19  Ontario
#> 148   182            34893 2020-06-20  Ontario
#> 149   156            35049 2020-06-21  Ontario
#> 150   229            35278 2020-06-22  Ontario
#> 151   238            35516 2020-06-23  Ontario
#> 152   202            35718 2020-06-24  Ontario
#> 153   178            35896 2020-06-25  Ontario
#> 154   119            36015 2020-06-26  Ontario
#> 155   166            36181 2020-06-27  Ontario
#> 156   253            36434 2020-06-28  Ontario
#> 157   237            36671 2020-06-29  Ontario
#> 158   151            36822 2020-06-30  Ontario
#> 159   125            36947 2020-07-01  Ontario
#> 160   158            37105 2020-07-02  Ontario
#> 161   165            37270 2020-07-03  Ontario
#> 162   118            37388 2020-07-04  Ontario
#> 163   126            37514 2020-07-05  Ontario
#> 164   178            37692 2020-07-06  Ontario
#> 165    84            37776 2020-07-07  Ontario
#> 166   164            37940 2020-07-08  Ontario
#> 167   124            38064 2020-07-09  Ontario
#> 168   117            38181 2020-07-10  Ontario
#> 169    81            38262 2020-07-11  Ontario
#> 170    68            38330 2020-07-12  Ontario
#> 171   254            38584 2020-07-13  Ontario
#> 172    83            38667 2020-07-14  Ontario
#> 173   104            38771 2020-07-15  Ontario
#> 174   117            38888 2020-07-16  Ontario
#> 175   142            39030 2020-07-17  Ontario
#> 176   205            39235 2020-07-18  Ontario
#> 177   150            39385 2020-07-19  Ontario
#> 178   143            39528 2020-07-20  Ontario
#> 179   189            39717 2020-07-21  Ontario
#> 180   142            39859 2020-07-22  Ontario
#> 181   131            39990 2020-07-23  Ontario
#> 182   215            40205 2020-07-24  Ontario
#> 183   131            40336 2020-07-25  Ontario
#> 184   107            40443 2020-07-26  Ontario
#> 185   144            40587 2020-07-27  Ontario
#> 186   123            40710 2020-07-28  Ontario
#> 187    97            40807 2020-07-29  Ontario
#> 188   101            40908 2020-07-30  Ontario
#> 189   117            41025 2020-07-31  Ontario
#> 190   113            41138 2020-08-01  Ontario
#> 191    85            41223 2020-08-02  Ontario
#> 192    80            41303 2020-08-03  Ontario
#> 193    94            41397 2020-08-04  Ontario
#> 194    78            41475 2020-08-05  Ontario
#> 195    89            41564 2020-08-06  Ontario
#> 196    87            41651 2020-08-07  Ontario
#> 197    89            41740 2020-08-08  Ontario
#> 198    61            41801 2020-08-09  Ontario
#> 199   121            41922 2020-08-10  Ontario
#> 200    96            42018 2020-08-11  Ontario
#> 201    74            42092 2020-08-12  Ontario
#> 202    70            42162 2020-08-13  Ontario
#> 203    85            42247 2020-08-14  Ontario
#> 204   101            42348 2020-08-15  Ontario
#> 205    53            42401 2020-08-16  Ontario
#> 206   126            42527 2020-08-17  Ontario
#> 207    99            42626 2020-08-18  Ontario
#> 208   105            42731 2020-08-19  Ontario
#> 209    93            42824 2020-08-20  Ontario
#> 210   145            42969 2020-08-21  Ontario
#> 211    79            43048 2020-08-22  Ontario
#> 212   104            43152 2020-08-23  Ontario
#> 213   145            43297 2020-08-24  Ontario
#> 214    90            43387 2020-08-25  Ontario
#> 215   125            43512 2020-08-26  Ontario
#> 216   133            43645 2020-08-27  Ontario
#> 217   117            43762 2020-08-28  Ontario
#> 218   120            43882 2020-08-29  Ontario
#> 219    94            43976 2020-08-30  Ontario
#> 220   152            44128 2020-08-31  Ontario
#> 221   135            44263 2020-09-01  Ontario
#> 222   137            44400 2020-09-02  Ontario
#> 223   154            44554 2020-09-03  Ontario
#> 224   122            44676 2020-09-04  Ontario
#> 225   160            44836 2020-09-05  Ontario
#> 226   126            44962 2020-09-06  Ontario
#> 227   186            45148 2020-09-07  Ontario
#> 228   240            45388 2020-09-08  Ontario
#> 229   179            45567 2020-09-09  Ontario
#> 230   167            45734 2020-09-10  Ontario
#> 231   234            45968 2020-09-11  Ontario
#> 232   185            46153 2020-09-12  Ontario
#> 233   224            46377 2020-09-13  Ontario
#> 234   370            46747 2020-09-14  Ontario
#> 235   257            47004 2020-09-15  Ontario
#> 236   328            47332 2020-09-16  Ontario
#> 237   338            47670 2020-09-17  Ontario
#> 238   402            48072 2020-09-18  Ontario
#> 239   380            48452 2020-09-19  Ontario
#> 240   294            48746 2020-09-20  Ontario
#> 241   501            49247 2020-09-21  Ontario
#> 242   525            49772 2020-09-22  Ontario
#> 243   346            50118 2020-09-23  Ontario
#> 244   393            50511 2020-09-24  Ontario
#> 245   486            50997 2020-09-25  Ontario
#> 246   478            51475 2020-09-26  Ontario
#> 247   452            51927 2020-09-27  Ontario
#> 248   760            52687 2020-09-28  Ontario
#> 249   561            53248 2020-09-29  Ontario
#> 250   643            53891 2020-09-30  Ontario
#> 251   588            54479 2020-10-01  Ontario
#> 252   726            55205 2020-10-02  Ontario
#> 253   656            55861 2020-10-03  Ontario
#> 254   495            56356 2020-10-04  Ontario
#> 255   691            57047 2020-10-05  Ontario
#> 256   567            57614 2020-10-06  Ontario
#> 257   647            58261 2020-10-07  Ontario
#> 258   835            59096 2020-10-08  Ontario
#> 259   906            60002 2020-10-09  Ontario
#> 260   876            60878 2020-10-10  Ontario
#> 261   571            61449 2020-10-11  Ontario
#> 262   755            62204 2020-10-12  Ontario
#> 263   902            63106 2020-10-13  Ontario
#> 264   762            63868 2020-10-14  Ontario
#> 265   716            64584 2020-10-15  Ontario
#> 266   913            65497 2020-10-16  Ontario
#> 267   634            66131 2020-10-17  Ontario
#> 268   631            66762 2020-10-18  Ontario
#> 269   752            67514 2020-10-19  Ontario
#> 270   793            68307 2020-10-20  Ontario
#> 271   813            69120 2020-10-21  Ontario
#> 272   886            70006 2020-10-22  Ontario
#> 273   860            70866 2020-10-23  Ontario
#> 274  1014            71880 2020-10-24  Ontario
#> 275   972            72852 2020-10-25  Ontario
#> 276   878            73730 2020-10-26  Ontario
#> 277   864            74594 2020-10-27  Ontario
#> 278   798            75392 2020-10-28  Ontario
#> 279   977            76369 2020-10-29  Ontario
#> 280  1010            77379 2020-10-30  Ontario
#> 281   868            78247 2020-10-31  Ontario
#> 282   943            79190 2020-11-01  Ontario
#> 283  1097            80287 2020-11-02  Ontario
#> 284  1123            81410 2020-11-03  Ontario
#> 285   990            82400 2020-11-04  Ontario
#> 286  1017            83417 2020-11-05  Ontario
#> 287  1056            84473 2020-11-06  Ontario
#> 288  1232            85705 2020-11-07  Ontario
#> 289  1173            86878 2020-11-08  Ontario
#> 290  1484            88362 2020-11-09  Ontario
#> 291  1549            89911 2020-11-10  Ontario
#> 292  1353            91264 2020-11-11  Ontario
#> 293  1582            92846 2020-11-12  Ontario
#> 294  1427            94273 2020-11-13  Ontario
#> 295  1371            95644 2020-11-14  Ontario
#> 296  1349            96993 2020-11-15  Ontario
#> 297  1459            98452 2020-11-16  Ontario
#> 298  1449            99901 2020-11-17  Ontario
#> 299  1280           101181 2020-11-18  Ontario
#> 300  1363           102544 2020-11-19  Ontario
#> 301  1462           104006 2020-11-20  Ontario
#> 302  1512           105518 2020-11-21  Ontario
#> 303  1423           106941 2020-11-22  Ontario
#> 304  1408           108349 2020-11-23  Ontario
#> 305  1394           109743 2020-11-24  Ontario
#> 306  1469           111212 2020-11-25  Ontario
#> 307  1567           112779 2020-11-26  Ontario
#> 308  1849           114628 2020-11-27  Ontario
#> 309  1715           116343 2020-11-28  Ontario
#> 310  1620           117963 2020-11-29  Ontario
#> 311  1775           119738 2020-11-30  Ontario
#> 312  2033           121771 2020-12-01  Ontario
#> 313  1815           123586 2020-12-02  Ontario
#> 314  1543           125129 2020-12-03  Ontario
#> 315  1902           127031 2020-12-04  Ontario
#> 316  1786           128817 2020-12-05  Ontario
#> 317  1971           130788 2020-12-06  Ontario
#> 318  1934           132722 2020-12-07  Ontario
#> 319  1887           134609 2020-12-08  Ontario
#> 320  1928           136537 2020-12-09  Ontario
#> 321  1929           138466 2020-12-10  Ontario
#> 322  1816           140282 2020-12-11  Ontario
#> 323  1785           142067 2020-12-12  Ontario
#> 324  1648           143715 2020-12-13  Ontario
#> 325  2328           146043 2020-12-14  Ontario
#> 326  2088           148131 2020-12-15  Ontario
#> 327  2250           150381 2020-12-16  Ontario
#> 328  2472           152853 2020-12-17  Ontario
#> 329  2286           155139 2020-12-18  Ontario
#> 330  2489           157628 2020-12-19  Ontario
#> 331  1955           159583 2020-12-20  Ontario
#> 332  2395           161978 2020-12-21  Ontario
#> 333  2272           164250 2020-12-22  Ontario
#> 334  2500           166750 2020-12-23  Ontario
#> 335  2358           169108 2020-12-24  Ontario
#> 336   441           169549 2020-12-25  Ontario
#> 337  1777           171326 2020-12-26  Ontario
#> 338  1961           173287 2020-12-27  Ontario
#> 339  3736           177023 2020-12-28  Ontario
#> 340  3059           180082 2020-12-29  Ontario
#> 341  2991           183073 2020-12-30  Ontario
#> 342  3244           186317 2020-12-31  Ontario
#> 343   739           187056 2020-01-01  Ontario
#> 344  3879           190935 2020-01-02  Ontario
#> 345  4171           195106 2020-01-03  Ontario
#> 346  3749           198855 2020-01-04  Ontario
#> 347  3268           202123 2020-01-05  Ontario
#> 348  3047           205170 2020-01-06  Ontario
#> 349  3372           208542 2020-01-07  Ontario
#> 350  4253           212795 2020-01-08  Ontario
#> 351  3757           216552 2020-01-09  Ontario
#> 352  3429           219981 2020-01-10  Ontario
#> 353  3883           223864 2020-01-11  Ontario
#> 354  2696           226560 2020-01-12  Ontario
#> 355  3271           229831 2020-01-13  Ontario
#> 356  3235           233066 2020-01-14  Ontario
#> 357  3056           236122 2020-01-15  Ontario
#> 358  2921           239043 2020-01-16  Ontario
#> 359  2949           241992 2020-01-17  Ontario
#> 360  2650           244642 2020-01-18  Ontario
#> 361  2388           247030 2020-01-19  Ontario
#> 362  2439           249469 2020-01-20  Ontario
#> 363  2908           252377 2020-01-21  Ontario
#> 364  2464           254841 2020-01-22  Ontario
#> 365  2231           257072 2020-01-23  Ontario
#> 366  1992           259064 2020-01-24  Ontario
#> 367  2170           261234 2020-01-25  Ontario
#> 368  1807           263041 2020-01-26  Ontario
#> 369  1759           264800 2020-01-27  Ontario
#> 370  2049           266849 2020-01-28  Ontario
#> 371  1955           268804 2020-01-29  Ontario
#> 372  1735           270539 2020-01-30  Ontario
#> 373  1826           272365 2020-01-31  Ontario
#> 374  1928           274293 2020-02-01  Ontario
#> 375   863           275156 2020-02-02  Ontario
#> 376  1189           276345 2020-02-03  Ontario
#> 377  1581           277926 2020-02-04  Ontario
#> 378  1641           279567 2020-02-05  Ontario
#> 379  1366           280933 2020-02-06  Ontario
#> 380  1368           282301 2020-02-07  Ontario
#> 381  1285           283586 2020-02-08  Ontario
#> 382   984           284570 2020-02-09  Ontario
#> 383  1131           285701 2020-02-10  Ontario
#> 384   967           286668 2020-02-11  Ontario
#> 385  1361           288029 2020-02-12  Ontario
#> 386  1276           289305 2020-02-13  Ontario
#> 387   582           289887 2020-02-14  Ontario
#> 388   987           290874 2020-02-15  Ontario
#> 389  1192           292066 2020-02-16  Ontario
#> 390   974           293040 2020-02-17  Ontario
#> 391  1117           294157 2020-02-18  Ontario
#> 392  1181           295338 2020-02-19  Ontario
#> 393  1106           296444 2020-02-20  Ontario
#> 394  1057           297501 2020-02-21  Ontario
#> 395  1182           298683 2020-02-22  Ontario
#> 396  1029           299712 2020-02-23  Ontario
#> 397  1098           300810 2020-02-24  Ontario
#> 398  1187           301997 2020-02-25  Ontario
#> 399  1294           303291 2020-02-26  Ontario
#> 400  1026           304317 2020-02-27  Ontario
#> 401   987           305304 2020-02-28  Ontario
#> 402  1202           306506 2020-03-01  Ontario
#> 403   933           307439 2020-03-02  Ontario
#> 404   921           308360 2020-03-03  Ontario
#> 405  1197           309557 2020-03-04  Ontario
#> 406  1111           310668 2020-03-05  Ontario
#> 407  1090           311758 2020-03-06  Ontario
#> 408  1352           313110 2020-03-07  Ontario
#> 409  1644           314754 2020-03-08  Ontario
#> 410  1295           316049 2020-03-09  Ontario
#> 411  1167           317216 2020-03-10  Ontario
#> 412  1349           318565 2020-03-11  Ontario
#> 413  1473           320038 2020-03-12  Ontario
#> 414  1521           321559 2020-03-13  Ontario
#> 415  1447           323006 2020-03-14  Ontario
#> 416  1344           324350 2020-03-15  Ontario
#> 417  1261           325611 2020-03-16  Ontario
#> 418  1570           327181 2020-03-17  Ontario
#> 419  1647           328828 2020-03-18  Ontario
#> 420  1870           330698 2020-03-19  Ontario
#> 421  1840           332538 2020-03-20  Ontario
#> 422  1612           334150 2020-03-21  Ontario
#> 423  1794           335944 2020-03-22  Ontario
#> 424  1636           337580 2020-03-23  Ontario
#> 425  1526           339106 2020-03-24  Ontario
#> 426  2451           341557 2020-03-25  Ontario
#> 427  2279           343836 2020-03-26  Ontario
#> 428  2499           346335 2020-03-27  Ontario
#> 429  2229           348564 2020-03-28  Ontario
#> 430  2294           350858 2020-03-29  Ontario
#> 431  2445           353303 2020-03-30  Ontario
#> 432  2585           355888 2020-03-31  Ontario
#> 433  2572           358460 2020-04-01  Ontario
#> 434  2549           361009 2020-04-02  Ontario
#> 435  3399           364408 2020-04-03  Ontario
#> 436  2673           367081 2020-04-04  Ontario
#> 437  2926           370007 2020-04-05  Ontario
#> 438  3944           373951 2020-04-06  Ontario
#> 439  3529           377480 2020-04-07  Ontario
#> 440  3526           381006 2020-04-08  Ontario
#> 441  4158           385164 2020-04-09  Ontario
#> 442  3823           388987 2020-04-10  Ontario
#> 443  4380           393367 2020-04-11  Ontario
#> 444  4638           398005 2020-04-12  Ontario
#> 445  3487           401492 2020-04-13  Ontario
#> 446  4708           406200 2020-04-14  Ontario
#> 447  4803           411003 2020-04-15  Ontario
#> 448  4810           415813 2020-04-16  Ontario
#> 449  3802           419615 2020-04-17  Ontario
#> 450  4387           424002 2020-04-18  Ontario
#> 451  4699           428701 2020-04-19  Ontario
#> 452  3474           432175 2020-04-20  Ontario
#> 453  3915           436090 2020-04-21  Ontario
#> 454  4408           440498 2020-04-22  Ontario
#> 455  4340           444838 2020-04-23  Ontario
#> 456  3817           448655 2020-04-24  Ontario
#> 457  3614           452269 2020-04-25  Ontario
#> 458  3793           456062 2020-04-26  Ontario
#> 459  3216           459278 2020-04-27  Ontario
#> 460  3843           463121 2020-04-28  Ontario
#> 461  5047           468168 2020-04-29  Ontario
#> 462  3695           471863 2020-04-30  Ontario
#> 463  3222           475085 2020-05-01  Ontario
#> 464  3507           478592 2020-05-02  Ontario
#> 465  2946           481538 2020-05-03  Ontario
#> 466  2807           484345 2020-05-04  Ontario
#> 467  3386           487731 2020-05-05  Ontario
#> 468  3269           491000 2020-05-06  Ontario
#> 469  3004           494004 2020-05-07  Ontario
#> 470  2726           496730 2020-05-08  Ontario
#> 471  2945           499675 2020-05-09  Ontario
#> 472  2627           502302 2020-05-10  Ontario
#> 473  2060           504362 2020-05-11  Ontario
#> 474  2693           507055 2020-05-12  Ontario
#> 475  2504           509559 2020-05-13  Ontario
#> 476  2543           512102 2020-05-14  Ontario
#> 477  2406           514508 2020-05-15  Ontario
#> 478  2039           516547 2020-05-16  Ontario
#> 479  1909           518456 2020-05-17  Ontario
#> 480  1411           519867 2020-05-18  Ontario
#> 481  2372           522239 2020-05-19  Ontario
#> 482  2059           524298 2020-05-20  Ontario
#> 483  1714           526012 2020-05-21  Ontario
#> 484  1712           527724 2020-05-22  Ontario
#> 485  1554           529278 2020-05-23  Ontario
#> 486   518           529796 2020-05-24  Ontario
#> 487  1885           531681 2020-05-25  Ontario
#> 488  1109           532790 2020-05-26  Ontario
#> 489  1150           533940 2020-05-27  Ontario
#> 490  1160           535100 2020-05-28  Ontario
#> 491   993           536093 2020-05-29  Ontario
#> 492   903           536996 2020-05-30  Ontario
#> 493   986           537982 2020-05-31  Ontario
#> 494   669           538651 2020-06-01  Ontario
#> 495   938           539589 2020-06-02  Ontario
#> 496   890           540479 2020-06-03  Ontario
#> 497   912           541391 2020-06-04  Ontario
#> 498   472           541863 2020-06-05  Ontario
#> 499   630           542493 2020-06-06  Ontario
#> 500   575           543068 2020-06-07  Ontario
#> 501   412           543480 2020-06-08  Ontario
#> 502   540           544020 2020-06-09  Ontario
#> 503   558           544578 2020-06-10  Ontario
#> 504   530           545108 2020-06-11  Ontario
#> 505   410           545518 2020-06-12  Ontario
#> 506   402           545920 2020-06-13  Ontario
#> 507   524           546444 2020-06-14  Ontario
#> 508   331           546775 2020-06-15  Ontario
#> 509   435           547210 2020-06-16  Ontario
#> 510   361           547571 2020-06-17  Ontario
#> 511   337           547908 2020-06-18  Ontario
#> 512   293           548201 2020-06-19  Ontario
#> 513   240           548441 2020-06-20  Ontario
#> 514   467           548908 2020-06-21  Ontario
#> 515   216           549124 2020-06-22  Ontario
#> 516   262           549386 2020-06-23  Ontario
#> 517   305           549691 2020-06-24  Ontario
#> 518   298           549989 2020-06-25  Ontario
#> 519   259           550248 2020-06-26  Ontario
#> 520   198           550446 2020-06-27  Ontario
#> 521   419           550865 2020-06-28  Ontario
#> 522   152           551017 2020-06-29  Ontario
#> 523   265           551282 2020-06-30  Ontario
#> 524    81           551363 2020-07-01  Ontario
#> 525   368           551731 2020-07-02  Ontario
#> 526   158           551889 2020-07-03  Ontario
#> 527   146           552035 2020-07-04  Ontario
#> 528   364           552399 2020-07-05  Ontario
#> 529   118           552517 2020-07-06  Ontario
#> 530   218           552735 2020-07-07  Ontario
#> 531   212           552947 2020-07-08  Ontario
#> 532   178           553125 2020-07-09  Ontario
#> 533   102           553227 2020-07-10  Ontario
#> 534    83           553310 2020-07-11  Ontario
#> 535   251           553561 2020-07-12  Ontario
#> 536   127           553688 2020-07-13  Ontario
#> 537   145           553833 2020-07-14  Ontario
#> 538   143           553976 2020-07-15  Ontario
#> 539   179           554155 2020-07-16  Ontario
#> 540   128           554283 2020-07-17  Ontario
#> 541   123           554406 2020-07-18  Ontario
#> 542   216           554622 2020-07-19  Ontario
#> 543   136           554758 2020-07-20  Ontario
#> 544   156           554914 2020-07-21  Ontario
#> 545   186           555100 2020-07-22  Ontario
#> 546   173           555273 2020-07-23  Ontario
#> 547   113           555386 2020-07-24  Ontario
#> 548   113           555499 2020-07-25  Ontario
#> 549   245           555744 2020-07-26  Ontario
#> 550   141           555885 2020-07-27  Ontario
#> 551   190           556075 2020-07-28  Ontario
#> 552   222           556297 2020-07-29  Ontario
#> 553   258           556555 2020-07-30  Ontario
#> 554   101           556656 2020-07-31  Ontario
#> 555    57           556713 2020-08-01  Ontario
#> 556    36           556749 2020-08-02  Ontario
#> 557   548           557297 2020-08-03  Ontario
#> 558   166           557463 2020-08-04  Ontario
#> 559   298           557761 2020-08-05  Ontario
#> 560   395           558156 2020-08-06  Ontario
#> 561   193           558349 2020-08-07  Ontario
#> 562   213           558562 2020-08-08  Ontario
#> 563   686           559248 2020-08-09  Ontario
#> 564   322           559570 2020-08-10  Ontario
#> 565   426           559996 2020-08-11  Ontario
#> 566   513           560509 2020-08-12  Ontario
#> 567   559           561068 2020-08-13  Ontario
#> 568   315           561383 2020-08-14  Ontario
#> 569   263           561646 2020-08-15  Ontario
#> 570   941           562587 2020-08-16  Ontario
#> 571   405           562992 2020-08-17  Ontario
#> 572   604           563596 2020-08-18  Ontario
#> 573   616           564212 2020-08-19  Ontario
#> 574   668           564880 2020-08-20  Ontario
#> 575   153           565033 2020-08-21  Ontario
#> 576   167           565200 2020-08-22  Ontario
#> 577  1585           566785 2020-08-23  Ontario
#> 578   534           567319 2020-08-24  Ontario
#> 579   637           567956 2020-08-25  Ontario
#> 580   813           568769 2020-08-26  Ontario
#> 581   829           569598 2020-08-27  Ontario
#> 582   333           569931 2020-08-28  Ontario
#> 583   218           570149 2020-08-29  Ontario
#> 584  1566           571715 2020-08-30  Ontario
#> 585   590           572305 2020-08-31  Ontario
#> 586   805           573110 2020-09-01  Ontario
#> 587   831           573941 2020-09-02  Ontario
#> 588   924           574865 2020-09-03  Ontario
#> 589   429           575294 2020-09-04  Ontario
#> 590   251           575545 2020-09-05  Ontario
#> 591   153           575698 2020-09-06  Ontario
#> 592  1906           577604 2020-09-07  Ontario
#> 593   627           578231 2020-09-08  Ontario
#> 594   845           579076 2020-09-09  Ontario
#> 595   890           579966 2020-09-10  Ontario
#> 596   466           580432 2020-09-11  Ontario
#> 597   276           580708 2020-09-12  Ontario
#> 598  1390           582098 2020-09-13  Ontario
#> 599   528           582626 2020-09-14  Ontario
#> 600   767           583393 2020-09-15  Ontario
#> 601   831           584224 2020-09-16  Ontario
#> 602   823           585047 2020-09-17  Ontario
#> 603   456           585503 2020-09-18  Ontario
#> 604   416           585919 2020-09-19  Ontario
#> 605  1121           587040 2020-09-20  Ontario
#> 606   472           587512 2020-09-21  Ontario
#> 607   651           588163 2020-09-22  Ontario
#> 608   671           588834 2020-09-23  Ontario
#> 609   698           589532 2020-09-24  Ontario
#> 610   380           589912 2020-09-25  Ontario
#> 611   269           590181 2020-09-26  Ontario
#> 612  1210           591391 2020-09-27  Ontario
#> 613   457           591848 2020-09-28  Ontario
#> 614   555           592403 2020-09-29  Ontario
#> 615   422           592825 2020-09-30  Ontario
#> 616   935           593760 2020-10-01  Ontario
#> 617   279           594039 2020-10-02  Ontario
#> 618   304           594343 2020-10-03  Ontario
#> 619  1137           595480 2020-10-04  Ontario
#> 620   419           595899 2020-10-05  Ontario
#> 621   562           596461 2020-10-06  Ontario
#> 622   586           597047 2020-10-07  Ontario
#> 623   638           597685 2020-10-08  Ontario
#> 624   220           597905 2020-10-09  Ontario
#> 625   186           598091 2020-10-10  Ontario
#> 626   163           598254 2020-10-11  Ontario
#> 627  1277           599531 2020-10-12  Ontario
#> 628   333           599864 2020-10-13  Ontario
#> 629   494           600358 2020-10-14  Ontario
#> 630   502           600860 2020-10-15  Ontario
#> 631   199           601059 2020-10-16  Ontario
#> 632   174           601233 2020-10-17  Ontario
#> 633   870           602103 2020-10-18  Ontario
#> 634   271           602374 2020-10-19  Ontario
#> 635   391           602765 2020-10-20  Ontario
#> 636   423           603188 2020-10-21  Ontario
#> 637   434           603622 2020-10-22  Ontario
#> 638   168           603790 2020-10-23  Ontario
#> 639   111           603901 2020-10-24  Ontario
#> 640   717           604618 2020-10-25  Ontario
#> 641   335           604953 2020-10-26  Ontario
#> 642   349           605302 2020-10-27  Ontario
#> 643   429           605731 2020-10-28  Ontario
#> 644   399           606130 2020-10-29  Ontario
#> 645   131           606261 2020-10-30  Ontario
#> 646   157           606418 2020-10-31  Ontario
#> 647   840           607258 2020-11-01  Ontario
#> 648   333           607591 2020-11-02  Ontario
#> 649   439           608030 2020-11-03  Ontario
#> 650   476           608506 2020-11-04  Ontario
#> 651   551           609057 2020-11-05  Ontario
#> 652   185           609242 2020-11-06  Ontario
#> 653   200           609442 2020-11-07  Ontario
#> 654  1223           610665 2020-11-08  Ontario
#> 655   426           611091 2020-11-09  Ontario
#> 656   602           611693 2020-11-10  Ontario
#> 657   308           612001 2020-11-11  Ontario
#> 658   942           612943 2020-11-12  Ontario
#> 659   232           613175 2020-11-13  Ontario
#> 660   217           613392 2020-11-14  Ontario
#> 661  1289           614681 2020-11-15  Ontario
#> 662   457           615138 2020-11-16  Ontario
#> 663   686           615824 2020-11-17  Ontario
#> 664   733           616557 2020-11-18  Ontario
#> 665   809           617366 2020-11-19  Ontario
#> 666   214           617580 2020-11-20  Ontario
#> 667   251           617831 2020-11-21  Ontario
#> 668  1502           619333 2020-11-22  Ontario
#> 669   608           619941 2020-11-23  Ontario
#> 670   705           620646 2020-11-24  Ontario
#> 671   792           621438 2020-11-25  Ontario
#> 672   921           622359 2020-11-26  Ontario
#> 673   255           622614 2020-11-27  Ontario
#> 674   273           622887 2020-11-28  Ontario
#> 675  1943           624830 2020-11-29  Ontario
#> 676   722           625552 2020-11-30  Ontario
#> 677   915           626467 2020-12-01  Ontario
#> 678   989           627456 2020-12-02  Ontario
#> 679  1112           628568 2020-12-03  Ontario
#> 680   310           628878 2020-12-04  Ontario
#> 681   290           629168 2020-12-05  Ontario
#> 682  2419           631587 2020-12-06  Ontario
#> 683   968           632555 2020-12-07  Ontario
#> 684  1258           633813 2020-12-08  Ontario
#> 685  1229           635042 2020-12-09  Ontario
#> 686  1623           636665 2020-12-10  Ontario
#> 687   452           637117 2020-12-11  Ontario
#> 688   382           637499 2020-12-12  Ontario
#> 689  3774           641273 2020-12-13  Ontario
#> 690  1564           642837 2020-12-14  Ontario
#> 691  2200           645037 2020-12-15  Ontario
#> 692  2947           647984 2020-12-16  Ontario
#> 693  3236           651220 2020-12-17  Ontario
#> 694  1008           652228 2020-12-18  Ontario
#> 695  1212           653440 2020-12-19  Ontario
#> 696  9156           662596 2020-12-20  Ontario
#> 697  4039           666635 2020-12-21  Ontario
#> 698  5121           671756 2020-12-22  Ontario
#> 699  7969           679725 2020-12-23  Ontario
#> 700  9393           689118 2020-12-24  Ontario
#> 701   535           689653 2020-12-25  Ontario
#> 702   974           690627 2020-12-26  Ontario
#> 703 12775           703402 2020-12-27  Ontario
#> 704  5015           708417 2020-12-28  Ontario
#> 705 32612           741029 2020-12-29  Ontario
#> 706 15430           756459 2020-12-30  Ontario
#> 707 17210           773669 2020-12-31  Ontario
get_cases(loc = "prov", date = "2021-09-01")
#> No encoding supplied: defaulting to UTF-8.
#>    cases cumulative_cases       date      province
#> 1   1315           254245 2020-09-01       Alberta
#> 2    785           166853 2020-09-01            BC
#> 3     49            58759 2020-09-01      Manitoba
#> 4     11             2774 2020-09-01 New Brunswick
#> 5      8             1486 2020-09-01            NL
#> 6      7             6038 2020-09-01   Nova Scotia
#> 7      0              657 2020-09-01       Nunavut
#> 8      0              461 2020-09-01           NWT
#> 9    805           573110 2020-09-01       Ontario
#> 10     3              233 2020-09-01           PEI
#> 11   690           389914 2020-09-01        Quebec
#> 12     0               13 2020-09-01   Repatriated
#> 13   321            54688 2020-09-01  Saskatchewan
#> 14     2              693 2020-09-01         Yukon