Query total cumulative deaths with ability to specify location and date range of returned data.

Query total cumulative deaths with ability to specify location and date range of returned data.

get_deaths(
  loc = "prov",
  date = NULL,
  after = "2020-01-01",
  before = Sys.Date(),
  datetime_type = TRUE
)

Arguments

loc

string: Specify geographic filter and aggregation of returned data. Valid loc arguments are: 'canada', 'prov' and two-letter province codes (e.g. 'ON', 'BC', etc.)

date

string: If not None, return data from the specified date YYYY-MM-DD. Superceeds 'after' and 'before' parameters.

after

string: Return data on and after the specified date YYYY-MM-DD.

before

string: Return data on and before the specified date YYYY-MM-DD.

datetime_type

boolean: Return date column as a character (FALSE) or as a Date (TRUE).

Value

Dataframe containing content of API response.

Examples

get_deaths(loc = "ON", before = "2021-12-31")
#> No encoding supplied: defaulting to UTF-8.
#>     cumulative_deaths       date deaths province
#> 1                   0 2020-03-08      0  Ontario
#> 2                   0 2020-03-09      0  Ontario
#> 3                   0 2020-03-10      0  Ontario
#> 4                   1 2020-03-11      1  Ontario
#> 5                   1 2020-03-12      0  Ontario
#> 6                   1 2020-03-13      0  Ontario
#> 7                   1 2020-03-14      0  Ontario
#> 8                   1 2020-03-15      0  Ontario
#> 9                   1 2020-03-16      0  Ontario
#> 10                  1 2020-03-17      0  Ontario
#> 11                  1 2020-03-18      0  Ontario
#> 12                  2 2020-03-19      1  Ontario
#> 13                  2 2020-03-20      0  Ontario
#> 14                  3 2020-03-21      1  Ontario
#> 15                  6 2020-03-22      3  Ontario
#> 16                  6 2020-03-23      0  Ontario
#> 17                  8 2020-03-24      2  Ontario
#> 18                 13 2020-03-25      5  Ontario
#> 19                 15 2020-03-26      2  Ontario
#> 20                 18 2020-03-27      3  Ontario
#> 21                 19 2020-03-28      1  Ontario
#> 22                 23 2020-03-29      4  Ontario
#> 23                 34 2020-03-30     11  Ontario
#> 24                 58 2020-03-31     24  Ontario
#> 25                 65 2020-04-01      7  Ontario
#> 26                 85 2020-04-02     20  Ontario
#> 27                102 2020-04-03     17  Ontario
#> 28                131 2020-04-04     29  Ontario
#> 29                142 2020-04-05     11  Ontario
#> 30                162 2020-04-06     20  Ontario
#> 31                188 2020-04-07     26  Ontario
#> 32                211 2020-04-08     23  Ontario
#> 33                231 2020-04-09     20  Ontario
#> 34                266 2020-04-10     35  Ontario
#> 35                283 2020-04-11     17  Ontario
#> 36                293 2020-04-12     10  Ontario
#> 37                329 2020-04-13     36  Ontario
#> 38                394 2020-04-14     65  Ontario
#> 39                442 2020-04-15     48  Ontario
#> 40                496 2020-04-16     54  Ontario
#> 41                529 2020-04-17     33  Ontario
#> 42                567 2020-04-18     38  Ontario
#> 43                601 2020-04-19     34  Ontario
#> 44                645 2020-04-20     44  Ontario
#> 45                693 2020-04-21     48  Ontario
#> 46                754 2020-04-22     61  Ontario
#> 47                795 2020-04-23     41  Ontario
#> 48                847 2020-04-24     52  Ontario
#> 49                895 2020-04-25     48  Ontario
#> 50                942 2020-04-26     47  Ontario
#> 51                999 2020-04-27     57  Ontario
#> 52               1047 2020-04-28     48  Ontario
#> 53               1130 2020-04-29     83  Ontario
#> 54               1173 2020-04-30     43  Ontario
#> 55               1234 2020-05-01     61  Ontario
#> 56               1261 2020-05-02     27  Ontario
#> 57               1302 2020-05-03     41  Ontario
#> 58               1422 2020-05-04    120  Ontario
#> 59               1484 2020-05-05     62  Ontario
#> 60               1545 2020-05-06     61  Ontario
#> 61               1592 2020-05-07     47  Ontario
#> 62               1645 2020-05-08     53  Ontario
#> 63               1709 2020-05-09     64  Ontario
#> 64               1727 2020-05-10     18  Ontario
#> 65               1773 2020-05-11     46  Ontario
#> 66               1838 2020-05-12     65  Ontario
#> 67               1871 2020-05-13     33  Ontario
#> 68               1898 2020-05-14     27  Ontario
#> 69               1925 2020-05-15     27  Ontario
#> 70               1959 2020-05-16     34  Ontario
#> 71               1982 2020-05-17     23  Ontario
#> 72               2002 2020-05-18     20  Ontario
#> 73               2017 2020-05-19     15  Ontario
#> 74               2065 2020-05-20     48  Ontario
#> 75               2090 2020-05-21     25  Ontario
#> 76               2112 2020-05-22     22  Ontario
#> 77               2139 2020-05-23     27  Ontario
#> 78               2163 2020-05-24     24  Ontario
#> 79               2193 2020-05-25     30  Ontario
#> 80               2216 2020-05-26     23  Ontario
#> 81               2247 2020-05-27     31  Ontario
#> 82               2275 2020-05-28     28  Ontario
#> 83               2297 2020-05-29     22  Ontario
#> 84               2316 2020-05-30     19  Ontario
#> 85               2328 2020-05-31     12  Ontario
#> 86               2336 2020-06-01      8  Ontario
#> 87               2355 2020-06-02     19  Ontario
#> 88               2373 2020-06-03     18  Ontario
#> 89               2418 2020-06-04     45  Ontario
#> 90               2428 2020-06-05     10  Ontario
#> 91               2465 2020-06-06     37  Ontario
#> 92               2484 2020-06-07     19  Ontario
#> 93               2507 2020-06-08     23  Ontario
#> 94               2519 2020-06-09     12  Ontario
#> 95               2528 2020-06-10      9  Ontario
#> 96               2545 2020-06-11     17  Ontario
#> 97               2556 2020-06-12     11  Ontario
#> 98               2564 2020-06-13      8  Ontario
#> 99               2573 2020-06-14      9  Ontario
#> 100              2580 2020-06-15      7  Ontario
#> 101              2594 2020-06-16     14  Ontario
#> 102              2601 2020-06-17      7  Ontario
#> 103              2606 2020-06-18      5  Ontario
#> 104              2616 2020-06-19     10  Ontario
#> 105              2646 2020-06-20     30  Ontario
#> 106              2651 2020-06-21      5  Ontario
#> 107              2659 2020-06-22      8  Ontario
#> 108              2669 2020-06-23     10  Ontario
#> 109              2683 2020-06-24     14  Ontario
#> 110              2695 2020-06-25     12  Ontario
#> 111              2698 2020-06-26      3  Ontario
#> 112              2707 2020-06-27      9  Ontario
#> 113              2712 2020-06-28      5  Ontario
#> 114              2722 2020-06-29     10  Ontario
#> 115              2726 2020-06-30      4  Ontario
#> 116              2730 2020-07-01      4  Ontario
#> 117              2734 2020-07-02      4  Ontario
#> 118              2737 2020-07-03      3  Ontario
#> 119              2739 2020-07-04      2  Ontario
#> 120              2740 2020-07-05      1  Ontario
#> 121              2740 2020-07-06      0  Ontario
#> 122              2741 2020-07-07      1  Ontario
#> 123              2746 2020-07-08      5  Ontario
#> 124              2748 2020-07-09      2  Ontario
#> 125              2758 2020-07-10     10  Ontario
#> 126              2758 2020-07-11      0  Ontario
#> 127              2759 2020-07-12      1  Ontario
#> 128              2765 2020-07-13      6  Ontario
#> 129              2767 2020-07-14      2  Ontario
#> 130              2776 2020-07-15      9  Ontario
#> 131              2782 2020-07-16      6  Ontario
#> 132              2787 2020-07-17      5  Ontario
#> 133              2788 2020-07-18      1  Ontario
#> 134              2791 2020-07-19      3  Ontario
#> 135              2791 2020-07-20      0  Ontario
#> 136              2794 2020-07-21      3  Ontario
#> 137              2795 2020-07-22      1  Ontario
#> 138              2795 2020-07-23      0  Ontario
#> 139              2798 2020-07-24      3  Ontario
#> 140              2800 2020-07-25      2  Ontario
#> 141              2804 2020-07-26      4  Ontario
#> 142              2805 2020-07-27      1  Ontario
#> 143              2810 2020-07-28      5  Ontario
#> 144              2812 2020-07-29      2  Ontario
#> 145              2815 2020-07-30      3  Ontario
#> 146              2817 2020-07-31      2  Ontario
#> 147              2819 2020-08-01      2  Ontario
#> 148              2820 2020-08-02      1  Ontario
#> 149              2823 2020-08-03      3  Ontario
#> 150              2826 2020-08-04      3  Ontario
#> 151              2826 2020-08-05      0  Ontario
#> 152              2827 2020-08-06      1  Ontario
#> 153              2827 2020-08-07      0  Ontario
#> 154              2829 2020-08-08      2  Ontario
#> 155              2830 2020-08-09      1  Ontario
#> 156              2830 2020-08-10      0  Ontario
#> 157              2831 2020-08-11      1  Ontario
#> 158              2831 2020-08-12      0  Ontario
#> 159              2832 2020-08-13      1  Ontario
#> 160              2833 2020-08-14      1  Ontario
#> 161              2833 2020-08-15      0  Ontario
#> 162              2833 2020-08-16      0  Ontario
#> 163              2833 2020-08-17      0  Ontario
#> 164              2834 2020-08-18      1  Ontario
#> 165              2834 2020-08-19      0  Ontario
#> 166              2835 2020-08-20      1  Ontario
#> 167              2839 2020-08-21      4  Ontario
#> 168              2839 2020-08-22      0  Ontario
#> 169              2840 2020-08-23      1  Ontario
#> 170              2840 2020-08-24      0  Ontario
#> 171              2843 2020-08-25      3  Ontario
#> 172              2844 2020-08-26      1  Ontario
#> 173              2844 2020-08-27      0  Ontario
#> 174              2851 2020-08-28      7  Ontario
#> 175              2852 2020-08-29      1  Ontario
#> 176              2852 2020-08-30      0  Ontario
#> 177              2853 2020-08-31      1  Ontario
#> 178              2854 2020-09-01      1  Ontario
#> 179              2855 2020-09-02      1  Ontario
#> 180              2856 2020-09-03      1  Ontario
#> 181              2856 2020-09-04      0  Ontario
#> 182              2856 2020-09-05      0  Ontario
#> 183              2856 2020-09-06      0  Ontario
#> 184              2857 2020-09-07      1  Ontario
#> 185              2858 2020-09-08      1  Ontario
#> 186              2859 2020-09-09      1  Ontario
#> 187              2860 2020-09-10      1  Ontario
#> 188              2860 2020-09-11      0  Ontario
#> 189              2861 2020-09-12      1  Ontario
#> 190              2862 2020-09-13      1  Ontario
#> 191              2863 2020-09-14      1  Ontario
#> 192              2867 2020-09-15      4  Ontario
#> 193              2869 2020-09-16      2  Ontario
#> 194              2869 2020-09-17      0  Ontario
#> 195              2870 2020-09-18      1  Ontario
#> 196              2871 2020-09-19      1  Ontario
#> 197              2871 2020-09-20      0  Ontario
#> 198              2874 2020-09-21      3  Ontario
#> 199              2878 2020-09-22      4  Ontario
#> 200              2881 2020-09-23      3  Ontario
#> 201              2882 2020-09-24      1  Ontario
#> 202              2882 2020-09-25      0  Ontario
#> 203              2883 2020-09-26      1  Ontario
#> 204              2884 2020-09-27      1  Ontario
#> 205              2885 2020-09-28      1  Ontario
#> 206              2889 2020-09-29      4  Ontario
#> 207              2893 2020-09-30      4  Ontario
#> 208              2896 2020-10-01      3  Ontario
#> 209              2980 2020-10-02     84  Ontario
#> 210              3020 2020-10-03     40  Ontario
#> 211              3023 2020-10-04      3  Ontario
#> 212              3032 2020-10-05      9  Ontario
#> 213              3036 2020-10-06      4  Ontario
#> 214              3037 2020-10-07      1  Ontario
#> 215              3040 2020-10-08      3  Ontario
#> 216              3045 2020-10-09      5  Ontario
#> 217              3052 2020-10-10      7  Ontario
#> 218              3054 2020-10-11      2  Ontario
#> 219              3056 2020-10-12      2  Ontario
#> 220              3065 2020-10-13      9  Ontario
#> 221              3067 2020-10-14      2  Ontario
#> 222              3076 2020-10-15      9  Ontario
#> 223              3086 2020-10-16     10  Ontario
#> 224              3090 2020-10-17      4  Ontario
#> 225              3095 2020-10-18      5  Ontario
#> 226              3099 2020-10-19      4  Ontario
#> 227              3103 2020-10-20      4  Ontario
#> 228              3112 2020-10-21      9  Ontario
#> 229              3121 2020-10-22      9  Ontario
#> 230              3128 2020-10-23      7  Ontario
#> 231              3131 2020-10-24      3  Ontario
#> 232              3138 2020-10-25      7  Ontario
#> 233              3144 2020-10-26      6  Ontario
#> 234              3149 2020-10-27      5  Ontario
#> 235              3156 2020-10-28      7  Ontario
#> 236              3166 2020-10-29     10  Ontario
#> 237              3177 2020-10-30     11  Ontario
#> 238              3183 2020-10-31      6  Ontario
#> 239              3191 2020-11-01      8  Ontario
#> 240              3201 2020-11-02     10  Ontario
#> 241              3217 2020-11-03     16  Ontario
#> 242              3232 2020-11-04     15  Ontario
#> 243              3247 2020-11-05     15  Ontario
#> 244              3264 2020-11-06     17  Ontario
#> 245              3274 2020-11-07     10  Ontario
#> 246              3282 2020-11-08      8  Ontario
#> 247              3298 2020-11-09     16  Ontario
#> 248              3314 2020-11-10     16  Ontario
#> 249              3336 2020-11-11     22  Ontario
#> 250              3350 2020-11-12     14  Ontario
#> 251              3366 2020-11-13     16  Ontario
#> 252              3383 2020-11-14     17  Ontario
#> 253              3402 2020-11-15     19  Ontario
#> 254              3412 2020-11-16     10  Ontario
#> 255              3428 2020-11-17     16  Ontario
#> 256              3460 2020-11-18     32  Ontario
#> 257              3484 2020-11-19     24  Ontario
#> 258              3496 2020-11-20     12  Ontario
#> 259              3516 2020-11-21     20  Ontario
#> 260              3529 2020-11-22     13  Ontario
#> 261              3545 2020-11-23     16  Ontario
#> 262              3567 2020-11-24     22  Ontario
#> 263              3598 2020-11-25     31  Ontario
#> 264              3617 2020-11-26     19  Ontario
#> 265              3643 2020-11-27     26  Ontario
#> 266              3666 2020-11-28     23  Ontario
#> 267              3686 2020-11-29     20  Ontario
#> 268              3699 2020-11-30     13  Ontario
#> 269              3714 2020-12-01     15  Ontario
#> 270              3747 2020-12-02     33  Ontario
#> 271              3762 2020-12-03     15  Ontario
#> 272              3794 2020-12-04     32  Ontario
#> 273              3808 2020-12-05     14  Ontario
#> 274              3835 2020-12-06     27  Ontario
#> 275              3853 2020-12-07     18  Ontario
#> 276              3869 2020-12-08     16  Ontario
#> 277              3904 2020-12-09     35  Ontario
#> 278              3939 2020-12-10     35  Ontario
#> 279              3964 2020-12-11     25  Ontario
#> 280              3987 2020-12-12     23  Ontario
#> 281              4002 2020-12-13     15  Ontario
#> 282              4024 2020-12-14     22  Ontario
#> 283              4051 2020-12-15     27  Ontario
#> 284              4085 2020-12-16     34  Ontario
#> 285              4113 2020-12-17     28  Ontario
#> 286              4144 2020-12-18     31  Ontario
#> 287              4184 2020-12-19     40  Ontario
#> 288              4198 2020-12-20     14  Ontario
#> 289              4222 2020-12-21     24  Ontario
#> 290              4253 2020-12-22     31  Ontario
#> 291              4300 2020-12-23     47  Ontario
#> 292              4349 2020-12-24     49  Ontario
#> 293              4350 2020-12-25      1  Ontario
#> 294              4359 2020-12-26      9  Ontario
#> 295              4375 2020-12-27     16  Ontario
#> 296              4490 2020-12-28    115  Ontario
#> 297              4518 2020-12-29     28  Ontario
#> 298              4562 2020-12-30     44  Ontario
#> 299              4606 2020-12-31     44  Ontario
#> 300              4619 2020-01-01     13  Ontario
#> 301              4655 2020-01-02     36  Ontario
#> 302              4697 2020-01-03     42  Ontario
#> 303              4738 2020-01-04     41  Ontario
#> 304              4793 2020-01-05     55  Ontario
#> 305              4865 2020-01-06     72  Ontario
#> 306              4921 2020-01-07     56  Ontario
#> 307              4974 2020-01-08     53  Ontario
#> 308              5019 2020-01-09     45  Ontario
#> 309              5046 2020-01-10     27  Ontario
#> 310              5090 2020-01-11     44  Ontario
#> 311              5150 2020-01-12     60  Ontario
#> 312              5222 2020-01-13     72  Ontario
#> 313              5277 2020-01-14     55  Ontario
#> 314              5335 2020-01-15     58  Ontario
#> 315              5393 2020-01-16     58  Ontario
#> 316              5461 2020-01-17     68  Ontario
#> 317              5494 2020-01-18     33  Ontario
#> 318              5548 2020-01-19     54  Ontario
#> 319              5635 2020-01-20     87  Ontario
#> 320              5688 2020-01-21     53  Ontario
#> 321              5756 2020-01-22     68  Ontario
#> 322              5798 2020-01-23     42  Ontario
#> 323              5836 2020-01-24     38  Ontario
#> 324              5902 2020-01-25     66  Ontario
#> 325              5957 2020-01-26     55  Ontario
#> 326              6007 2020-01-27     50  Ontario
#> 327              6054 2020-01-28     47  Ontario
#> 328              6128 2020-01-29     74  Ontario
#> 329              6187 2020-01-30     59  Ontario
#> 330              6213 2020-01-31     26  Ontario
#> 331              6257 2020-02-01     44  Ontario
#> 332              6297 2020-02-02     40  Ontario
#> 333              6345 2020-02-03     48  Ontario
#> 334              6402 2020-02-04     57  Ontario
#> 335              6454 2020-02-05     52  Ontario
#> 336              6498 2020-02-06     44  Ontario
#> 337              6513 2020-02-07     15  Ontario
#> 338              6544 2020-02-08     31  Ontario
#> 339              6580 2020-02-09     36  Ontario
#> 340              6611 2020-02-10     31  Ontario
#> 341              6639 2020-02-11     28  Ontario
#> 342              6658 2020-02-12     19  Ontario
#> 343              6682 2020-02-13     24  Ontario
#> 344              6688 2020-02-14      6  Ontario
#> 345              6694 2020-02-15      6  Ontario
#> 346              6735 2020-02-16     41  Ontario
#> 347              6750 2020-02-17     15  Ontario
#> 348              6806 2020-02-18     56  Ontario
#> 349              6844 2020-02-19     38  Ontario
#> 350              6864 2020-02-20     20  Ontario
#> 351              6874 2020-02-21     10  Ontario
#> 352              6883 2020-02-22      9  Ontario
#> 353              6900 2020-02-23     17  Ontario
#> 354              6912 2020-02-24     12  Ontario
#> 355              6932 2020-02-25     20  Ontario
#> 356              6958 2020-02-26     26  Ontario
#> 357              6978 2020-02-27     20  Ontario
#> 358              6990 2020-02-28     12  Ontario
#> 359              6999 2020-03-01      9  Ontario
#> 360              7012 2020-03-02     13  Ontario
#> 361              7027 2020-03-03     15  Ontario
#> 362              7041 2020-03-04     14  Ontario
#> 363              7057 2020-03-05     16  Ontario
#> 364              7063 2020-03-06      6  Ontario
#> 365              7073 2020-03-07     10  Ontario
#> 366              7083 2020-03-08     10  Ontario
#> 367              7096 2020-03-09     13  Ontario
#> 368              7109 2020-03-10     13  Ontario
#> 369              7121 2020-03-11     12  Ontario
#> 370              7138 2020-03-12     17  Ontario
#> 371              7146 2020-03-13      8  Ontario
#> 372              7158 2020-03-14     12  Ontario
#> 373              7168 2020-03-15     10  Ontario
#> 374              7180 2020-03-16     12  Ontario
#> 375              7193 2020-03-17     13  Ontario
#> 376              7205 2020-03-18     12  Ontario
#> 377              7216 2020-03-19     11  Ontario
#> 378              7231 2020-03-20     15  Ontario
#> 379              7241 2020-03-21     10  Ontario
#> 380              7248 2020-03-22      7  Ontario
#> 381              7259 2020-03-23     11  Ontario
#> 382              7271 2020-03-24     12  Ontario
#> 383              7288 2020-03-25     17  Ontario
#> 384              7297 2020-03-26      9  Ontario
#> 385              7308 2020-03-27     11  Ontario
#> 386              7322 2020-03-28     14  Ontario
#> 387              7332 2020-03-29     10  Ontario
#> 388              7351 2020-03-30     19  Ontario
#> 389              7370 2020-03-31     19  Ontario
#> 390              7402 2020-04-01     32  Ontario
#> 391              7416 2020-04-02     14  Ontario
#> 392              7428 2020-04-03     12  Ontario
#> 393              7439 2020-04-04     11  Ontario
#> 394              7446 2020-04-05      7  Ontario
#> 395              7463 2020-04-06     17  Ontario
#> 396              7483 2020-04-07     20  Ontario
#> 397              7497 2020-04-08     14  Ontario
#> 398              7512 2020-04-09     15  Ontario
#> 399              7529 2020-04-10     17  Ontario
#> 400              7546 2020-04-11     17  Ontario
#> 401              7561 2020-04-12     15  Ontario
#> 402              7589 2020-04-13     28  Ontario
#> 403              7615 2020-04-14     26  Ontario
#> 404              7645 2020-04-15     30  Ontario
#> 405              7672 2020-04-16     27  Ontario
#> 406              7700 2020-04-17     28  Ontario
#> 407              7724 2020-04-18     24  Ontario
#> 408              7739 2020-04-19     15  Ontario
#> 409              7766 2020-04-20     27  Ontario
#> 410              7812 2020-04-21     46  Ontario
#> 411              7835 2020-04-22     23  Ontario
#> 412              7867 2020-04-23     32  Ontario
#> 413              7888 2020-04-24     21  Ontario
#> 414              7912 2020-04-25     24  Ontario
#> 415              7939 2020-04-26     27  Ontario
#> 416              7967 2020-04-27     28  Ontario
#> 417              8007 2020-04-28     40  Ontario
#> 418              8048 2020-04-29     41  Ontario
#> 419              8073 2020-04-30     25  Ontario
#> 420              8083 2020-05-01     10  Ontario
#> 421              8090 2020-05-02      7  Ontario
#> 422              8131 2020-05-03     41  Ontario
#> 423              8167 2020-05-04     36  Ontario
#> 424              8200 2020-05-05     33  Ontario
#> 425              8226 2020-05-06     26  Ontario
#> 426              8249 2020-05-07     23  Ontario
#> 427              8265 2020-05-08     16  Ontario
#> 428              8278 2020-05-09     13  Ontario
#> 429              8322 2020-05-10     44  Ontario
#> 430              8344 2020-05-11     22  Ontario
#> 431              8379 2020-05-12     35  Ontario
#> 432              8416 2020-05-13     37  Ontario
#> 433              8440 2020-05-14     24  Ontario
#> 434              8452 2020-05-15     12  Ontario
#> 435              8456 2020-05-16      4  Ontario
#> 436              8490 2020-05-17     34  Ontario
#> 437              8507 2020-05-18     17  Ontario
#> 438              8527 2020-05-19     20  Ontario
#> 439              8555 2020-05-20     28  Ontario
#> 440              8586 2020-05-21     31  Ontario
#> 441              8593 2020-05-22      7  Ontario
#> 442              8603 2020-05-23     10  Ontario
#> 443              8607 2020-05-24      4  Ontario
#> 444              8653 2020-05-25     46  Ontario
#> 445              8681 2020-05-26     28  Ontario
#> 446              8697 2020-05-27     16  Ontario
#> 447              8713 2020-05-28     16  Ontario
#> 448              8720 2020-05-29      7  Ontario
#> 449              8728 2020-05-30      8  Ontario
#> 450              8753 2020-05-31     25  Ontario
#> 451              8765 2020-06-01     12  Ontario
#> 452              8789 2020-06-02     24  Ontario
#> 453              8804 2020-06-03     15  Ontario
#> 454              8827 2020-06-04     23  Ontario
#> 455              8835 2020-06-05      8  Ontario
#> 456              8841 2020-06-06      6  Ontario
#> 457              8868 2020-06-07     27  Ontario
#> 458              8889 2020-06-08     21  Ontario
#> 459              8911 2020-06-09     22  Ontario
#> 460              8924 2020-06-10     13  Ontario
#> 461              8931 2020-06-11      7  Ontario
#> 462              8937 2020-06-12      6  Ontario
#> 463              8937 2020-06-13      0  Ontario
#> 464              8952 2020-06-14     15  Ontario
#> 465              8962 2020-06-15     10  Ontario
#> 466              8992 2020-06-16     30  Ontario
#> 467              9000 2020-06-17      8  Ontario
#> 468              9012 2020-06-18     12  Ontario
#> 469              9016 2020-06-19      4  Ontario
#> 470              9019 2020-06-20      3  Ontario
#> 471              9055 2020-06-21     36  Ontario
#> 472              9069 2020-06-22     14  Ontario
#> 473              9078 2020-06-23      9  Ontario
#> 474              9086 2020-06-24      8  Ontario
#> 475              9096 2020-06-25     10  Ontario
#> 476              9101 2020-06-26      5  Ontario
#> 477              9102 2020-06-27      1  Ontario
#> 478              9121 2020-06-28     19  Ontario
#> 479              9142 2020-06-29     21  Ontario
#> 480              9163 2020-06-30     21  Ontario
#> 481              9165 2020-07-01      2  Ontario
#> 482              9180 2020-07-02     15  Ontario
#> 483              9183 2020-07-03      3  Ontario
#> 484              9185 2020-07-04      2  Ontario
#> 485              9198 2020-07-05     13  Ontario
#> 486              9206 2020-07-06      8  Ontario
#> 487              9208 2020-07-07      2  Ontario
#> 488              9213 2020-07-08      5  Ontario
#> 489              9218 2020-07-09      5  Ontario
#> 490              9220 2020-07-10      2  Ontario
#> 491              9221 2020-07-11      1  Ontario
#> 492              9232 2020-07-12     11  Ontario
#> 493              9238 2020-07-13      6  Ontario
#> 494              9249 2020-07-14     11  Ontario
#> 495              9257 2020-07-15      8  Ontario
#> 496              9261 2020-07-16      4  Ontario
#> 497              9262 2020-07-17      1  Ontario
#> 498              9264 2020-07-18      2  Ontario
#> 499              9271 2020-07-19      7  Ontario
#> 500              9273 2020-07-20      2  Ontario
#> 501              9277 2020-07-21      4  Ontario
#> 502              9279 2020-07-22      2  Ontario
#> 503              9282 2020-07-23      3  Ontario
#> 504              9286 2020-07-24      4  Ontario
#> 505              9287 2020-07-25      1  Ontario
#> 506              9289 2020-07-26      2  Ontario
#> 507              9295 2020-07-27      6  Ontario
#> 508              9300 2020-07-28      5  Ontario
#> 509              9311 2020-07-29     11  Ontario
#> 510              9317 2020-07-30      6  Ontario
#> 511              9322 2020-07-31      5  Ontario
#> 512              9323 2020-08-01      1  Ontario
#> 513              9323 2020-08-02      0  Ontario
#> 514              9327 2020-08-03      4  Ontario
#> 515              9339 2020-08-04     12  Ontario
#> 516              9352 2020-08-05     13  Ontario
#> 517              9365 2020-08-06     13  Ontario
#> 518              9367 2020-08-07      2  Ontario
#> 519              9367 2020-08-08      0  Ontario
#> 520              9373 2020-08-09      6  Ontario
#> 521              9374 2020-08-10      1  Ontario
#> 522              9378 2020-08-11      4  Ontario
#> 523              9381 2020-08-12      3  Ontario
#> 524              9385 2020-08-13      4  Ontario
#> 525              9387 2020-08-14      2  Ontario
#> 526              9387 2020-08-15      0  Ontario
#> 527              9391 2020-08-16      4  Ontario
#> 528              9400 2020-08-17      9  Ontario
#> 529              9412 2020-08-18     12  Ontario
#> 530              9421 2020-08-19      9  Ontario
#> 531              9432 2020-08-20     11  Ontario
#> 532              9429 2020-08-21     -3  Ontario
#> 533              9429 2020-08-22      0  Ontario
#> 534              9432 2020-08-23      3  Ontario
#> 535              9440 2020-08-24      8  Ontario
#> 536              9450 2020-08-25     10  Ontario
#> 537              9452 2020-08-26      2  Ontario
#> 538              9458 2020-08-27      6  Ontario
#> 539              9464 2020-08-28      6  Ontario
#> 540              9464 2020-08-29      0  Ontario
#> 541              9467 2020-08-30      3  Ontario
#> 542              9477 2020-08-31     10  Ontario
#> 543              9488 2020-09-01     11  Ontario
#> 544              9498 2020-09-02     10  Ontario
#> 545              9506 2020-09-03      8  Ontario
#> 546              9511 2020-09-04      5  Ontario
#> 547              9512 2020-09-05      1  Ontario
#> 548              9513 2020-09-06      1  Ontario
#> 549              9524 2020-09-07     11  Ontario
#> 550              9540 2020-09-08     16  Ontario
#> 551              9552 2020-09-09     12  Ontario
#> 552              9565 2020-09-10     13  Ontario
#> 553              9566 2020-09-11      1  Ontario
#> 554              9570 2020-09-12      4  Ontario
#> 555              9581 2020-09-13     11  Ontario
#> 556              9591 2020-09-14     10  Ontario
#> 557              9593 2020-09-15      2  Ontario
#> 558              9598 2020-09-16      5  Ontario
#> 559              9606 2020-09-17      8  Ontario
#> 560              9607 2020-09-18      1  Ontario
#> 561              9610 2020-09-19      3  Ontario
#> 562              9620 2020-09-20     10  Ontario
#> 563              9630 2020-09-21     10  Ontario
#> 564              9637 2020-09-22      7  Ontario
#> 565              9643 2020-09-23      6  Ontario
#> 566              9653 2020-09-24     10  Ontario
#> 567              9659 2020-09-25      6  Ontario
#> 568              9659 2020-09-26      0  Ontario
#> 569              9665 2020-09-27      6  Ontario
#> 570              9678 2020-09-28     13  Ontario
#> 571              9688 2020-09-29     10  Ontario
#> 572              9695 2020-09-30      7  Ontario
#> 573              9708 2020-10-01     13  Ontario
#> 574              9709 2020-10-02      1  Ontario
#> 575              9710 2020-10-03      1  Ontario
#> 576              9716 2020-10-04      6  Ontario
#> 577              9727 2020-10-05     11  Ontario
#> 578              9734 2020-10-06      7  Ontario
#> 579              9740 2020-10-07      6  Ontario
#> 580              9746 2020-10-08      6  Ontario
#> 581              9746 2020-10-09      0  Ontario
#> 582              9748 2020-10-10      2  Ontario
#> 583              9748 2020-10-11      0  Ontario
#> 584              9753 2020-10-12      5  Ontario
#> 585              9765 2020-10-13     12  Ontario
#> 586              9768 2020-10-14      3  Ontario
#> 587              9772 2020-10-15      4  Ontario
#> 588              9773 2020-10-16      1  Ontario
#> 589              9773 2020-10-17      0  Ontario
#> 590              9777 2020-10-18      4  Ontario
#> 591              9782 2020-10-19      5  Ontario
#> 592              9786 2020-10-20      4  Ontario
#> 593              9794 2020-10-21      8  Ontario
#> 594              9803 2020-10-22      9  Ontario
#> 595              9804 2020-10-23      1  Ontario
#> 596              9805 2020-10-24      1  Ontario
#> 597              9807 2020-10-25      2  Ontario
#> 598              9812 2020-10-26      5  Ontario
#> 599              9823 2020-10-27     11  Ontario
#> 600              9826 2020-10-28      3  Ontario
#> 601              9827 2020-10-29      1  Ontario
#> 602              9828 2020-10-30      1  Ontario
#> 603              9829 2020-10-31      1  Ontario
#> 604              9836 2020-11-01      7  Ontario
#> 605              9844 2020-11-02      8  Ontario
#> 606              9849 2020-11-03      5  Ontario
#> 607              9852 2020-11-04      3  Ontario
#> 608              9855 2020-11-05      3  Ontario
#> 609              9857 2020-11-06      2  Ontario
#> 610              9858 2020-11-07      1  Ontario
#> 611              9857 2020-11-08     -1  Ontario
#> 612              9864 2020-11-09      7  Ontario
#> 613              9869 2020-11-10      5  Ontario
#> 614              9871 2020-11-11      2  Ontario
#> 615              9878 2020-11-12      7  Ontario
#> 616              9881 2020-11-13      3  Ontario
#> 617              9882 2020-11-14      1  Ontario
#> 618              9893 2020-11-15     11  Ontario
#> 619              9899 2020-11-16      6  Ontario
#> 620              9909 2020-11-17     10  Ontario
#> 621              9913 2020-11-18      4  Ontario
#> 622              9918 2020-11-19      5  Ontario
#> 623              9919 2020-11-20      1  Ontario
#> 624              9920 2020-11-21      1  Ontario
#> 625              9924 2020-11-22      4  Ontario
#> 626              9939 2020-11-23     15  Ontario
#> 627              9946 2020-11-24      7  Ontario
#> 628              9949 2020-11-25      3  Ontario
#> 629              9953 2020-11-26      4  Ontario
#> 630              9955 2020-11-27      2  Ontario
#> 631              9955 2020-11-28      0  Ontario
#> 632              9958 2020-11-29      3  Ontario
#> 633              9963 2020-11-30      5  Ontario
#> 634              9965 2020-12-01      2  Ontario
#> 635              9972 2020-12-02      7  Ontario
#> 636              9980 2020-12-03      8  Ontario
#> 637              9980 2020-12-04      0  Ontario
#> 638              9981 2020-12-05      1  Ontario
#> 639              9988 2020-12-06      7  Ontario
#> 640             10001 2020-12-07     13  Ontario
#> 641             10010 2020-12-08      9  Ontario
#> 642             10013 2020-12-09      3  Ontario
#> 643             10024 2020-12-10     11  Ontario
#> 644             10024 2020-12-11      0  Ontario
#> 645             10025 2020-12-12      1  Ontario
#> 646             10034 2020-12-13      9  Ontario
#> 647             10044 2020-12-14     10  Ontario
#> 648             10052 2020-12-15      8  Ontario
#> 649             10058 2020-12-16      6  Ontario
#> 650             10063 2020-12-17      5  Ontario
#> 651             10063 2020-12-18      0  Ontario
#> 652             10063 2020-12-19      0  Ontario
#> 653             10065 2020-12-20      2  Ontario
#> 654             10083 2020-12-21     18  Ontario
#> 655             10091 2020-12-22      8  Ontario
#> 656             10097 2020-12-23      6  Ontario
#> 657             10103 2020-12-24      6  Ontario
#> 658             10104 2020-12-25      1  Ontario
#> 659             10104 2020-12-26      0  Ontario
#> 660             10111 2020-12-27      7  Ontario
#> 661             10117 2020-12-28      6  Ontario
#> 662             10124 2020-12-29      7  Ontario
#> 663             10139 2020-12-30     15  Ontario
#> 664             10152 2020-12-31     13  Ontario
get_deaths(loc = "prov", date = "2021-09-01")
#> No encoding supplied: defaulting to UTF-8.
#>    cumulative_deaths       date deaths      province
#> 1               2383 2020-09-01      8       Alberta
#> 2               1818 2020-09-01      2            BC
#> 3               1190 2020-09-01      1      Manitoba
#> 4                 46 2020-09-01      0 New Brunswick
#> 5                  7 2020-09-01      0            NL
#> 6                 94 2020-09-01      0   Nova Scotia
#> 7                  4 2020-09-01      0       Nunavut
#> 8                  1 2020-09-01      0           NWT
#> 9               9488 2020-09-01     11       Ontario
#> 10                 0 2020-09-01      0           PEI
#> 11             11288 2020-09-01      2        Quebec
#> 12                 0 2020-09-01      0   Repatriated
#> 13               607 2020-09-01      4  Saskatchewan
#> 14                 8 2020-09-01      0         Yukon