Query total cumulative recovered cases with ability to specify location and date range of returned data.

Query total cumulative recovered cases with ability to specify location and date range of returned data.

get_recoveries(
  loc = "prov",
  date = NULL,
  after = "2020-01-01",
  before = Sys.Date(),
  datetime_type = TRUE
)

Arguments

loc

string: Specify geographic filter and aggregation of returned data. Valid loc arguments are: 'canada', 'prov' and two-letter province codes (e.g. 'ON', 'BC', etc.)

date

string: If not None, return data from the specified date YYYY-MM-DD. Superceeds 'after' and 'before' parameters.

after

string: Return data on and after the specified date YYYY-MM-DD.

before

string: Return data on and before the specified date YYYY-MM-DD.

datetime_type

boolean: Return date column as a character (FALSE) or as a Date (TRUE).

Value

Dataframe containing content of API response.

Examples

get_recoveries(loc = "ON", before = "2021-12-31")
#> No encoding supplied: defaulting to UTF-8.
#>     cumulative_recovered       date province recovered
#> 1                      1 2020-02-12  Ontario         1
#> 2                      1 2020-02-13  Ontario         0
#> 3                      1 2020-02-14  Ontario         0
#> 4                      1 2020-02-15  Ontario         0
#> 5                      1 2020-02-16  Ontario         0
#> 6                      1 2020-02-17  Ontario         0
#> 7                      1 2020-02-18  Ontario         0
#> 8                      1 2020-02-19  Ontario         0
#> 9                      2 2020-02-20  Ontario         1
#> 10                     3 2020-02-21  Ontario         1
#> 11                     3 2020-02-22  Ontario         0
#> 12                     3 2020-02-23  Ontario         0
#> 13                     3 2020-02-24  Ontario         0
#> 14                     3 2020-02-25  Ontario         0
#> 15                     3 2020-02-26  Ontario         0
#> 16                     3 2020-02-27  Ontario         0
#> 17                     3 2020-02-28  Ontario         0
#> 18                     3 2020-02-29  Ontario         0
#> 19                     3 2020-03-01  Ontario         0
#> 20                     3 2020-03-02  Ontario         0
#> 21                     3 2020-03-03  Ontario         0
#> 22                     3 2020-03-04  Ontario         0
#> 23                     4 2020-03-05  Ontario         1
#> 24                     4 2020-03-06  Ontario         0
#> 25                     4 2020-03-07  Ontario         0
#> 26                     4 2020-03-08  Ontario         0
#> 27                     5 2020-03-09  Ontario         1
#> 28                     5 2020-03-10  Ontario         0
#> 29                     5 2020-03-11  Ontario         0
#> 30                     5 2020-03-12  Ontario         0
#> 31                     5 2020-03-13  Ontario         0
#> 32                     5 2020-03-14  Ontario         0
#> 33                     5 2020-03-15  Ontario         0
#> 34                     5 2020-03-16  Ontario         0
#> 35                     5 2020-03-17  Ontario         0
#> 36                     5 2020-03-18  Ontario         0
#> 37                     5 2020-03-19  Ontario         0
#> 38                     5 2020-03-20  Ontario         0
#> 39                     6 2020-03-21  Ontario         1
#> 40                     8 2020-03-22  Ontario         2
#> 41                     8 2020-03-23  Ontario         0
#> 42                     8 2020-03-24  Ontario         0
#> 43                     8 2020-03-25  Ontario         0
#> 44                     8 2020-03-26  Ontario         0
#> 45                     8 2020-03-27  Ontario         0
#> 46                     8 2020-03-28  Ontario         0
#> 47                     8 2020-03-29  Ontario         0
#> 48                   431 2020-03-30  Ontario       423
#> 49                   534 2020-03-31  Ontario       103
#> 50                   689 2020-04-01  Ontario       155
#> 51                   831 2020-04-02  Ontario       142
#> 52                  1023 2020-04-03  Ontario       192
#> 53                  1219 2020-04-04  Ontario       196
#> 54                  1449 2020-04-05  Ontario       230
#> 55                  1624 2020-04-06  Ontario       175
#> 56                  1802 2020-04-07  Ontario       178
#> 57                  2074 2020-04-08  Ontario       272
#> 58                  2305 2020-04-09  Ontario       231
#> 59                  2574 2020-04-10  Ontario       269
#> 60                  2858 2020-04-11  Ontario       284
#> 61                  3121 2020-04-12  Ontario       263
#> 62                  3357 2020-04-13  Ontario       236
#> 63                  3568 2020-04-14  Ontario       211
#> 64                  3902 2020-04-15  Ontario       334
#> 65                  4194 2020-04-16  Ontario       292
#> 66                  4556 2020-04-17  Ontario       362
#> 67                  4875 2020-04-18  Ontario       319
#> 68                  5209 2020-04-19  Ontario       334
#> 69                  5515 2020-04-20  Ontario       306
#> 70                  5806 2020-04-21  Ontario       291
#> 71                  6221 2020-04-22  Ontario       415
#> 72                  6680 2020-04-23  Ontario       459
#> 73                  7087 2020-04-24  Ontario       407
#> 74                  7509 2020-04-25  Ontario       422
#> 75                  8000 2020-04-26  Ontario       491
#> 76                  8525 2020-04-27  Ontario       525
#> 77                  8964 2020-04-28  Ontario       439
#> 78                  9612 2020-04-29  Ontario       648
#> 79                 10205 2020-04-30  Ontario       593
#> 80                 10825 2020-05-01  Ontario       620
#> 81                 11390 2020-05-02  Ontario       565
#> 82                 12005 2020-05-03  Ontario       615
#> 83                 12505 2020-05-04  Ontario       500
#> 84                 12779 2020-05-05  Ontario       274
#> 85                 13222 2020-05-06  Ontario       443
#> 86                 13569 2020-05-07  Ontario       347
#> 87                 13990 2020-05-08  Ontario       421
#> 88                 14383 2020-05-09  Ontario       393
#> 89                 14772 2020-05-10  Ontario       389
#> 90                 15131 2020-05-11  Ontario       359
#> 91                 15391 2020-05-12  Ontario       260
#> 92                 15845 2020-05-13  Ontario       454
#> 93                 16204 2020-05-14  Ontario       359
#> 94                 16641 2020-05-15  Ontario       437
#> 95                 17020 2020-05-16  Ontario       379
#> 96                 17360 2020-05-17  Ontario       340
#> 97                 17638 2020-05-18  Ontario       278
#> 98                 17898 2020-05-19  Ontario       260
#> 99                 18190 2020-05-20  Ontario       292
#> 100                18509 2020-05-21  Ontario       319
#> 101                18767 2020-05-22  Ontario       258
#> 102                19146 2020-05-23  Ontario       379
#> 103                19477 2020-05-24  Ontario       331
#> 104                19698 2020-05-25  Ontario       221
#> 105                19958 2020-05-26  Ontario       260
#> 106                20372 2020-05-27  Ontario       414
#> 107                20673 2020-05-28  Ontario       301
#> 108                20983 2020-05-29  Ontario       310
#> 109                21353 2020-05-30  Ontario       370
#> 110                21810 2020-05-31  Ontario       457
#> 111                22153 2020-06-01  Ontario       343
#> 112                22484 2020-06-02  Ontario       331
#> 113                22811 2020-06-03  Ontario       327
#> 114                23208 2020-06-04  Ontario       397
#> 115                23583 2020-06-05  Ontario       375
#> 116                23947 2020-06-06  Ontario       364
#> 117                24252 2020-06-07  Ontario       305
#> 118                24492 2020-06-08  Ontario       240
#> 119                24829 2020-06-09  Ontario       337
#> 120                25380 2020-06-10  Ontario       551
#> 121                25885 2020-06-11  Ontario       505
#> 122                26187 2020-06-12  Ontario       302
#> 123                26538 2020-06-13  Ontario       351
#> 124                26961 2020-06-14  Ontario       423
#> 125                27213 2020-06-15  Ontario       252
#> 126                27431 2020-06-16  Ontario       218
#> 127                27784 2020-06-17  Ontario       353
#> 128                28004 2020-06-18  Ontario       220
#> 129                28250 2020-06-19  Ontario       246
#> 130                28468 2020-06-20  Ontario       218
#> 131                28719 2020-06-21  Ontario       251
#> 132                28933 2020-06-22  Ontario       214
#> 133                29107 2020-06-23  Ontario       174
#> 134                29336 2020-06-24  Ontario       229
#> 135                29528 2020-06-25  Ontario       192
#> 136                29754 2020-06-26  Ontario       226
#> 137                29932 2020-06-27  Ontario       178
#> 138                30107 2020-06-28  Ontario       175
#> 139                30196 2020-06-29  Ontario        89
#> 140                30344 2020-06-30  Ontario       148
#> 141                30344 2020-07-01  Ontario         0
#> 142                30730 2020-07-02  Ontario       386
#> 143                30909 2020-07-03  Ontario       179
#> 144                31083 2020-07-04  Ontario       174
#> 145                31266 2020-07-05  Ontario       183
#> 146                31426 2020-07-06  Ontario       160
#> 147                31603 2020-07-07  Ontario       177
#> 148                31805 2020-07-08  Ontario       202
#> 149                31977 2020-07-09  Ontario       172
#> 150                32155 2020-07-10  Ontario       178
#> 151                32422 2020-07-11  Ontario       267
#> 152                32534 2020-07-12  Ontario       112
#> 153                32663 2020-07-13  Ontario       129
#> 154                32785 2020-07-14  Ontario       122
#> 155                32920 2020-07-15  Ontario       135
#> 156                33061 2020-07-16  Ontario       141
#> 157                33162 2020-07-17  Ontario       101
#> 158                33294 2020-07-18  Ontario       132
#> 159                33407 2020-07-19  Ontario       113
#> 160                33513 2020-07-20  Ontario       106
#> 161                33605 2020-07-21  Ontario        92
#> 162                33812 2020-07-22  Ontario       207
#> 163                33963 2020-07-23  Ontario       151
#> 164                34100 2020-07-24  Ontario       137
#> 165                34240 2020-07-25  Ontario       140
#> 166                34359 2020-07-26  Ontario       119
#> 167                34461 2020-07-27  Ontario       102
#> 168                34567 2020-07-28  Ontario       106
#> 169                34741 2020-07-29  Ontario       174
#> 170                34906 2020-07-30  Ontario       165
#> 171                35074 2020-07-31  Ontario       168
#> 172                35237 2020-08-01  Ontario       163
#> 173                35359 2020-08-02  Ontario       122
#> 174                35359 2020-08-03  Ontario         0
#> 175                35601 2020-08-04  Ontario       242
#> 176                35747 2020-08-05  Ontario       146
#> 177                35906 2020-08-06  Ontario       159
#> 178                36024 2020-08-07  Ontario       118
#> 179                36131 2020-08-08  Ontario       107
#> 180                36279 2020-08-09  Ontario       148
#> 181                36381 2020-08-10  Ontario       102
#> 182                36456 2020-08-11  Ontario        75
#> 183                36590 2020-08-12  Ontario       134
#> 184                36689 2020-08-13  Ontario        99
#> 185                36772 2020-08-14  Ontario        83
#> 186                36873 2020-08-15  Ontario       101
#> 187                36953 2020-08-16  Ontario        80
#> 188                37036 2020-08-17  Ontario        83
#> 189                37126 2020-08-18  Ontario        90
#> 190                37215 2020-08-19  Ontario        89
#> 191                37291 2020-08-20  Ontario        76
#> 192                37397 2020-08-21  Ontario       106
#> 193                37487 2020-08-22  Ontario        90
#> 194                37595 2020-08-23  Ontario       108
#> 195                37673 2020-08-24  Ontario        78
#> 196                37748 2020-08-25  Ontario        75
#> 197                37863 2020-08-26  Ontario       115
#> 198                37940 2020-08-27  Ontario        77
#> 199                38023 2020-08-28  Ontario        83
#> 200                38126 2020-08-29  Ontario       103
#> 201                38204 2020-08-30  Ontario        78
#> 202                38277 2020-08-31  Ontario        73
#> 203                38369 2020-09-01  Ontario        92
#> 204                38506 2020-09-02  Ontario       137
#> 205                38625 2020-09-03  Ontario       119
#> 206                38741 2020-09-04  Ontario       116
#> 207                38847 2020-09-05  Ontario       106
#> 208                38958 2020-09-06  Ontario       111
#> 209                38958 2020-09-07  Ontario         0
#> 210                39196 2020-09-08  Ontario       238
#> 211                39332 2020-09-09  Ontario       136
#> 212                39474 2020-09-10  Ontario       142
#> 213                39598 2020-09-11  Ontario       124
#> 214                39717 2020-09-12  Ontario       119
#> 215                39841 2020-09-13  Ontario       124
#> 216                39974 2020-09-14  Ontario       133
#> 217                40091 2020-09-15  Ontario       117
#> 218                40245 2020-09-16  Ontario       154
#> 219                40424 2020-09-17  Ontario       179
#> 220                40600 2020-09-18  Ontario       176
#> 221                40777 2020-09-19  Ontario       177
#> 222                40968 2020-09-20  Ontario       191
#> 223                41146 2020-09-21  Ontario       178
#> 224                41342 2020-09-22  Ontario       196
#> 225                41600 2020-09-23  Ontario       258
#> 226                41886 2020-09-24  Ontario       286
#> 227                42169 2020-09-25  Ontario       283
#> 228                42507 2020-09-26  Ontario       338
#> 229                42796 2020-09-27  Ontario       289
#> 230                43127 2020-09-28  Ontario       331
#> 231                43450 2020-09-29  Ontario       323
#> 232                43907 2020-09-30  Ontario       457
#> 233                44422 2020-10-01  Ontario       515
#> 234                44850 2020-10-02  Ontario       428
#> 235                45285 2020-10-03  Ontario       435
#> 236                45819 2020-10-04  Ontario       534
#> 237                46360 2020-10-05  Ontario       541
#> 238                46906 2020-10-06  Ontario       546
#> 239                47613 2020-10-07  Ontario       707
#> 240                48308 2020-10-08  Ontario       695
#> 241                49032 2020-10-09  Ontario       724
#> 242                49732 2020-10-10  Ontario       700
#> 243                50437 2020-10-11  Ontario       705
#> 244                50437 2020-10-12  Ontario         0
#> 245                51729 2020-10-13  Ontario      1292
#> 246                52512 2020-10-14  Ontario       783
#> 247                53291 2020-10-15  Ontario       779
#> 248                54004 2020-10-16  Ontario       713
#> 249                54686 2020-10-17  Ontario       682
#> 250                55371 2020-10-18  Ontario       685
#> 251                55978 2020-10-19  Ontario       607
#> 252                56606 2020-10-20  Ontario       628
#> 253                57325 2020-10-21  Ontario       719
#> 254                58066 2020-10-22  Ontario       741
#> 255                58799 2020-10-23  Ontario       733
#> 256                59424 2020-10-24  Ontario       625
#> 257                60160 2020-10-25  Ontario       736
#> 258                60839 2020-10-26  Ontario       679
#> 259                61530 2020-10-27  Ontario       691
#> 260                62303 2020-10-28  Ontario       773
#> 261                63123 2020-10-29  Ontario       820
#> 262                63919 2020-10-30  Ontario       796
#> 263                64717 2020-10-31  Ontario       798
#> 264                65581 2020-11-01  Ontario       864
#> 265                66407 2020-11-02  Ontario       826
#> 266                67244 2020-11-03  Ontario       837
#> 267                68189 2020-11-04  Ontario       945
#> 268                69137 2020-11-05  Ontario       948
#> 269                70086 2020-11-06  Ontario       949
#> 270                70938 2020-11-07  Ontario       852
#> 271                71815 2020-11-08  Ontario       877
#> 272                72636 2020-11-09  Ontario       821
#> 273                73417 2020-11-10  Ontario       781
#> 274                74303 2020-11-11  Ontario       886
#> 275                75220 2020-11-12  Ontario       917
#> 276                76238 2020-11-13  Ontario      1018
#> 277                77241 2020-11-14  Ontario      1003
#> 278                78303 2020-11-15  Ontario      1062
#> 279                79295 2020-11-16  Ontario       992
#> 280                80430 2020-11-17  Ontario      1135
#> 281                81925 2020-11-18  Ontario      1495
#> 282                83301 2020-11-19  Ontario      1376
#> 283                84716 2020-11-20  Ontario      1415
#> 284                86079 2020-11-21  Ontario      1363
#> 285                87508 2020-11-22  Ontario      1429
#> 286                88992 2020-11-23  Ontario      1484
#> 287                90074 2020-11-24  Ontario      1082
#> 288                91550 2020-11-25  Ontario      1476
#> 289                92915 2020-11-26  Ontario      1365
#> 290                94366 2020-11-27  Ontario      1451
#> 291                95876 2020-11-28  Ontario      1510
#> 292                97319 2020-11-29  Ontario      1443
#> 293                98639 2020-11-30  Ontario      1320
#> 294               100012 2020-12-01  Ontario      1373
#> 295               101698 2020-12-02  Ontario      1686
#> 296               103239 2020-12-03  Ontario      1541
#> 297               104792 2020-12-04  Ontario      1553
#> 298               106416 2020-12-05  Ontario      1624
#> 299               107990 2020-12-06  Ontario      1574
#> 300               109402 2020-12-07  Ontario      1412
#> 301               110951 2020-12-08  Ontario      1549
#> 302               112875 2020-12-09  Ontario      1924
#> 303               114679 2020-12-10  Ontario      1804
#> 304               116432 2020-12-11  Ontario      1753
#> 305               118350 2020-12-12  Ontario      1918
#> 306               120028 2020-12-13  Ontario      1678
#> 307               121563 2020-12-14  Ontario      1535
#> 308               123373 2020-12-15  Ontario      1810
#> 309               125416 2020-12-16  Ontario      2043
#> 310               127425 2020-12-17  Ontario      2009
#> 311               129417 2020-12-18  Ontario      1992
#> 312               131282 2020-12-19  Ontario      1865
#> 313               133213 2020-12-20  Ontario      1931
#> 314               134867 2020-12-21  Ontario      1654
#> 315               136767 2020-12-22  Ontario      1900
#> 316               139010 2020-12-23  Ontario      2243
#> 317               141023 2020-12-24  Ontario      2013
#> 318               141023 2020-12-25  Ontario         0
#> 319               145173 2020-12-26  Ontario      4150
#> 320               147178 2020-12-27  Ontario      2005
#> 321               149329 2020-12-28  Ontario      2151
#> 322               151562 2020-12-29  Ontario      2233
#> 323               153799 2020-12-30  Ontario      2237
#> 324               156012 2020-12-31  Ontario      2213
#> 325               158472 2020-01-01  Ontario      2460
#> 326               160526 2020-01-02  Ontario      2054
#> 327               162701 2020-01-03  Ontario      2175
#> 328               164775 2020-01-04  Ontario      2074
#> 329               166790 2020-01-05  Ontario      2015
#> 330               169795 2020-01-06  Ontario      3005
#> 331               172571 2020-01-07  Ontario      2776
#> 332               175309 2020-01-08  Ontario      2738
#> 333               178224 2020-01-09  Ontario      2915
#> 334               180720 2020-01-10  Ontario      2496
#> 335               183476 2020-01-11  Ontario      2756
#> 336               186829 2020-01-12  Ontario      3353
#> 337               190221 2020-01-13  Ontario      3392
#> 338               193814 2020-01-14  Ontario      3593
#> 339               197194 2020-01-15  Ontario      3380
#> 340               200406 2020-01-16  Ontario      3212
#> 341               203484 2020-01-17  Ontario      3078
#> 342               206310 2020-01-18  Ontario      2826
#> 343               209183 2020-01-19  Ontario      2873
#> 344               212897 2020-01-20  Ontario      3714
#> 345               215887 2020-01-21  Ontario      2990
#> 346               219262 2020-01-22  Ontario      3375
#> 347               222287 2020-01-23  Ontario      3025
#> 348               225046 2020-01-24  Ontario      2759
#> 349               227494 2020-01-25  Ontario      2448
#> 350               229755 2020-01-26  Ontario      2261
#> 351               232480 2020-01-27  Ontario      2725
#> 352               234971 2020-01-28  Ontario      2491
#> 353               237871 2020-01-29  Ontario      2900
#> 354               240494 2020-01-30  Ontario      2623
#> 355               242807 2020-01-31  Ontario      2313
#> 356               244939 2020-02-01  Ontario      2132
#> 357               247236 2020-02-02  Ontario      2297
#> 358               248981 2020-02-03  Ontario      1745
#> 359               250937 2020-02-04  Ontario      1956
#> 360               253170 2020-02-05  Ontario      2233
#> 361               254966 2020-02-06  Ontario      1796
#> 362               256903 2020-02-07  Ontario      1937
#> 363               258603 2020-02-08  Ontario      1700
#> 364               259991 2020-02-09  Ontario      1388
#> 365               261700 2020-02-10  Ontario      1709
#> 366               263044 2020-02-11  Ontario      1344
#> 367               264459 2020-02-12  Ontario      1415
#> 368               265893 2020-02-13  Ontario      1434
#> 369               267128 2020-02-14  Ontario      1235
#> 370               268301 2020-02-15  Ontario      1173
#> 371               269413 2020-02-16  Ontario      1112
#> 372               270869 2020-02-17  Ontario      1456
#> 373               272146 2020-02-18  Ontario      1277
#> 374               273401 2020-02-19  Ontario      1255
#> 375               274714 2020-02-20  Ontario      1313
#> 376               275854 2020-02-21  Ontario      1140
#> 377               276937 2020-02-22  Ontario      1083
#> 378               277939 2020-02-23  Ontario      1002
#> 379               279230 2020-02-24  Ontario      1291
#> 380               280324 2020-02-25  Ontario      1094
#> 381               281331 2020-02-26  Ontario      1007
#> 382               282315 2020-02-27  Ontario       984
#> 383               283344 2020-02-28  Ontario      1029
#> 384               284283 2020-03-01  Ontario       939
#> 385               285262 2020-03-02  Ontario       979
#> 386               286352 2020-03-03  Ontario      1090
#> 387               287424 2020-03-04  Ontario      1072
#> 388               288583 2020-03-05  Ontario      1159
#> 389               289735 2020-03-06  Ontario      1152
#> 390               290840 2020-03-07  Ontario      1105
#> 391               291834 2020-03-08  Ontario       994
#> 392               292806 2020-03-09  Ontario       972
#> 393               294018 2020-03-10  Ontario      1212
#> 394               295128 2020-03-11  Ontario      1110
#> 395               296252 2020-03-12  Ontario      1124
#> 396               297403 2020-03-13  Ontario      1151
#> 397               298570 2020-03-14  Ontario      1167
#> 398               299684 2020-03-15  Ontario      1114
#> 399               300769 2020-03-16  Ontario      1085
#> 400               302257 2020-03-17  Ontario      1488
#> 401               303493 2020-03-18  Ontario      1236
#> 402               304789 2020-03-19  Ontario      1296
#> 403               306050 2020-03-20  Ontario      1261
#> 404               307403 2020-03-21  Ontario      1353
#> 405               308578 2020-03-22  Ontario      1175
#> 406               309849 2020-03-23  Ontario      1271
#> 407               311380 2020-03-24  Ontario      1531
#> 408               312709 2020-03-25  Ontario      1329
#> 409               314384 2020-03-26  Ontario      1675
#> 410               315865 2020-03-27  Ontario      1481
#> 411               317408 2020-03-28  Ontario      1543
#> 412               318932 2020-03-29  Ontario      1524
#> 413               320409 2020-03-30  Ontario      1477
#> 414               322382 2020-03-31  Ontario      1973
#> 415               324196 2020-04-01  Ontario      1814
#> 416               326121 2020-04-02  Ontario      1925
#> 417               327940 2020-04-03  Ontario      1819
#> 418               327940 2020-04-04  Ontario         0
#> 419               331600 2020-04-05  Ontario      3660
#> 420               333576 2020-04-06  Ontario      1976
#> 421               335983 2020-04-07  Ontario      2407
#> 422               338559 2020-04-08  Ontario      2576
#> 423               341200 2020-04-09  Ontario      2641
#> 424               343622 2020-04-10  Ontario      2422
#> 425               346239 2020-04-11  Ontario      2617
#> 426               348684 2020-04-12  Ontario      2445
#> 427               351257 2020-04-13  Ontario      2573
#> 428               354417 2020-04-14  Ontario      3160
#> 429               357591 2020-04-15  Ontario      3174
#> 430               360742 2020-04-16  Ontario      3151
#> 431               364353 2020-04-17  Ontario      3611
#> 432               367691 2020-04-18  Ontario      3338
#> 433               370844 2020-04-19  Ontario      3153
#> 434               374213 2020-04-20  Ontario      3369
#> 435               378417 2020-04-21  Ontario      4204
#> 436               383014 2020-04-22  Ontario      4597
#> 437               387712 2020-04-23  Ontario      4698
#> 438               392044 2020-04-24  Ontario      4332
#> 439               396283 2020-04-25  Ontario      4239
#> 440               400340 2020-04-26  Ontario      4057
#> 441               404248 2020-04-27  Ontario      3908
#> 442               408606 2020-04-28  Ontario      4358
#> 443               413010 2020-04-29  Ontario      4404
#> 444               417252 2020-04-30  Ontario      4242
#> 445               421216 2020-05-01  Ontario      3964
#> 446               425163 2020-05-02  Ontario      3947
#> 447               428786 2020-05-03  Ontario      3623
#> 448               432109 2020-05-04  Ontario      3323
#> 449               436470 2020-05-05  Ontario      4361
#> 450               440467 2020-05-06  Ontario      3997
#> 451               444342 2020-05-07  Ontario      3875
#> 452               447938 2020-05-08  Ontario      3596
#> 453               451591 2020-05-09  Ontario      3653
#> 454               454701 2020-05-10  Ontario      3110
#> 455               457599 2020-05-11  Ontario      2898
#> 456               461076 2020-05-12  Ontario      3477
#> 457               464531 2020-05-13  Ontario      3455
#> 458               468033 2020-05-14  Ontario      3502
#> 459               471096 2020-05-15  Ontario      3063
#> 460               474175 2020-05-16  Ontario      3079
#> 461               477128 2020-05-17  Ontario      2953
#> 462               479630 2020-05-18  Ontario      2502
#> 463               482749 2020-05-19  Ontario      3119
#> 464               485512 2020-05-20  Ontario      2763
#> 465               488201 2020-05-21  Ontario      2689
#> 466               490721 2020-05-22  Ontario      2520
#> 467               493179 2020-05-23  Ontario      2458
#> 468               495251 2020-05-24  Ontario      2072
#> 469               497269 2020-05-25  Ontario      2018
#> 470               499640 2020-05-26  Ontario      2371
#> 471               501942 2020-05-27  Ontario      2302
#> 472               504304 2020-05-28  Ontario      2362
#> 473               506361 2020-05-29  Ontario      2057
#> 474               508428 2020-05-30  Ontario      2067
#> 475               510135 2020-05-31  Ontario      1707
#> 476               511703 2020-06-01  Ontario      1568
#> 477               513436 2020-06-02  Ontario      1733
#> 478               514999 2020-06-03  Ontario      1563
#> 479               516396 2020-06-04  Ontario      1397
#> 480               517638 2020-06-05  Ontario      1242
#> 481               518860 2020-06-06  Ontario      1222
#> 482               519801 2020-06-07  Ontario       941
#> 483               520811 2020-06-08  Ontario      1010
#> 484               521743 2020-06-09  Ontario       932
#> 485               522682 2020-06-10  Ontario       939
#> 486               523532 2020-06-11  Ontario       850
#> 487               524362 2020-06-12  Ontario       830
#> 488               525125 2020-06-13  Ontario       763
#> 489               525795 2020-06-14  Ontario       670
#> 490               526440 2020-06-15  Ontario       645
#> 491               527162 2020-06-16  Ontario       722
#> 492               527797 2020-06-17  Ontario       635
#> 493               528421 2020-06-18  Ontario       624
#> 494               529002 2020-06-19  Ontario       581
#> 495               529506 2020-06-20  Ontario       504
#> 496               529992 2020-06-21  Ontario       486
#> 497               530434 2020-06-22  Ontario       442
#> 498               530894 2020-06-23  Ontario       460
#> 499               531278 2020-06-24  Ontario       384
#> 500               531571 2020-06-25  Ontario       293
#> 501               532074 2020-06-26  Ontario       503
#> 502               532453 2020-06-27  Ontario       379
#> 503               532779 2020-06-28  Ontario       326
#> 504               533150 2020-06-29  Ontario       371
#> 505               533472 2020-06-30  Ontario       322
#> 506               533472 2020-07-01  Ontario         0
#> 507               534010 2020-07-02  Ontario       538
#> 508               534272 2020-07-03  Ontario       262
#> 509               534558 2020-07-04  Ontario       286
#> 510               534791 2020-07-05  Ontario       233
#> 511               535110 2020-07-06  Ontario       319
#> 512               535346 2020-07-07  Ontario       236
#> 513               535577 2020-07-08  Ontario       231
#> 514               535810 2020-07-09  Ontario       233
#> 515               536028 2020-07-10  Ontario       218
#> 516               536306 2020-07-11  Ontario       278
#> 517               536402 2020-07-12  Ontario        96
#> 518               536603 2020-07-13  Ontario       201
#> 519               536819 2020-07-14  Ontario       216
#> 520               536987 2020-07-15  Ontario       168
#> 521               537176 2020-07-16  Ontario       189
#> 522               537379 2020-07-17  Ontario       203
#> 523               537545 2020-07-18  Ontario       166
#> 524               537698 2020-07-19  Ontario       153
#> 525               537824 2020-07-20  Ontario       126
#> 526               537975 2020-07-21  Ontario       151
#> 527               538124 2020-07-22  Ontario       149
#> 528               538271 2020-07-23  Ontario       147
#> 529               538421 2020-07-24  Ontario       150
#> 530               538565 2020-07-25  Ontario       144
#> 531               538702 2020-07-26  Ontario       137
#> 532               538860 2020-07-27  Ontario       158
#> 533               539031 2020-07-28  Ontario       171
#> 534               539200 2020-07-29  Ontario       169
#> 535               539348 2020-07-30  Ontario       148
#> 536               539485 2020-07-31  Ontario       137
#> 537               539640 2020-08-01  Ontario       155
#> 538               539640 2020-08-02  Ontario         0
#> 539               539920 2020-08-03  Ontario       280
#> 540               540075 2020-08-04  Ontario       155
#> 541               540258 2020-08-05  Ontario       183
#> 542               540407 2020-08-06  Ontario       149
#> 543               540598 2020-08-07  Ontario       191
#> 544               540830 2020-08-08  Ontario       232
#> 545               541022 2020-08-09  Ontario       192
#> 546               541222 2020-08-10  Ontario       200
#> 547               541426 2020-08-11  Ontario       204
#> 548               541682 2020-08-12  Ontario       256
#> 549               541946 2020-08-13  Ontario       264
#> 550               542206 2020-08-14  Ontario       260
#> 551               542476 2020-08-15  Ontario       270
#> 552               542797 2020-08-16  Ontario       321
#> 553               543077 2020-08-17  Ontario       280
#> 554               543422 2020-08-18  Ontario       345
#> 555               543760 2020-08-19  Ontario       338
#> 556               544204 2020-08-20  Ontario       444
#> 557               544679 2020-08-21  Ontario       475
#> 558               545070 2020-08-22  Ontario       391
#> 559               545572 2020-08-23  Ontario       502
#> 560               546113 2020-08-24  Ontario       541
#> 561               546685 2020-08-25  Ontario       572
#> 562               547198 2020-08-26  Ontario       513
#> 563               547823 2020-08-27  Ontario       625
#> 564               548398 2020-08-28  Ontario       575
#> 565               548959 2020-08-29  Ontario       561
#> 566               549499 2020-08-30  Ontario       540
#> 567               550179 2020-08-31  Ontario       680
#> 568               550829 2020-09-01  Ontario       650
#> 569               551510 2020-09-02  Ontario       681
#> 570               552248 2020-09-03  Ontario       738
#> 571               552976 2020-09-04  Ontario       728
#> 572               553549 2020-09-05  Ontario       573
#> 573               553549 2020-09-06  Ontario         0
#> 574               562131 2020-09-07  Ontario      8582
#> 575               562936 2020-09-08  Ontario       805
#> 576               563676 2020-09-09  Ontario       740
#> 577               564441 2020-09-10  Ontario       765
#> 578               564708 2020-09-11  Ontario       267
#> 579               564949 2020-09-12  Ontario       241
#> 580               566532 2020-09-13  Ontario      1583
#> 581               567297 2020-09-14  Ontario       765
#> 582               567989 2020-09-15  Ontario       692
#> 583               568662 2020-09-16  Ontario       673
#> 584               569641 2020-09-17  Ontario       979
#> 585               569903 2020-09-18  Ontario       262
#> 586               570106 2020-09-19  Ontario       203
#> 587               571342 2020-09-20  Ontario      1236
#> 588               572155 2020-09-21  Ontario       813
#> 589               572878 2020-09-22  Ontario       723
#> 590               573668 2020-09-23  Ontario       790
#> 591               574457 2020-09-24  Ontario       789
#> 592               574747 2020-09-25  Ontario       290
#> 593               574980 2020-09-26  Ontario       233
#> 594               576492 2020-09-27  Ontario      1512
#> 595               577286 2020-09-28  Ontario       794
#> 596               577955 2020-09-29  Ontario       669
#> 597               578443 2020-09-30  Ontario       488
#> 598               579251 2020-10-01  Ontario       808
#> 599               579470 2020-10-02  Ontario       219
#> 600               579672 2020-10-03  Ontario       202
#> 601               580945 2020-10-04  Ontario      1273
#> 602               581664 2020-10-05  Ontario       719
#> 603               582242 2020-10-06  Ontario       578
#> 604               582845 2020-10-07  Ontario       603
#> 605               583445 2020-10-08  Ontario       600
#> 606               583724 2020-10-09  Ontario       279
#> 607               583930 2020-10-10  Ontario       206
#> 608               584101 2020-10-11  Ontario       171
#> 609               585659 2020-10-12  Ontario      1558
#> 610               586149 2020-10-13  Ontario       490
#> 611               586677 2020-10-14  Ontario       528
#> 612               587201 2020-10-15  Ontario       524
#> 613               587413 2020-10-16  Ontario       212
#> 614               587571 2020-10-17  Ontario       158
#> 615               588673 2020-10-18  Ontario      1102
#> 616               589239 2020-10-19  Ontario       566
#> 617               589710 2020-10-20  Ontario       471
#> 618               590112 2020-10-21  Ontario       402
#> 619               590512 2020-10-22  Ontario       400
#> 620               590694 2020-10-23  Ontario       182
#> 621               590831 2020-10-24  Ontario       137
#> 622               591740 2020-10-25  Ontario       909
#> 623               592183 2020-10-26  Ontario       443
#> 624               592527 2020-10-27  Ontario       344
#> 625               592899 2020-10-28  Ontario       372
#> 626               593268 2020-10-29  Ontario       369
#> 627               593428 2020-10-30  Ontario       160
#> 628               593544 2020-10-31  Ontario       116
#> 629               594279 2020-11-01  Ontario       735
#> 630               594704 2020-11-02  Ontario       425
#> 631               595056 2020-11-03  Ontario       352
#> 632               595379 2020-11-04  Ontario       323
#> 633               595733 2020-11-05  Ontario       354
#> 634               595865 2020-11-06  Ontario       132
#> 635               595975 2020-11-07  Ontario       110
#> 636               596828 2020-11-08  Ontario       853
#> 637               597321 2020-11-09  Ontario       493
#> 638               597721 2020-11-10  Ontario       400
#> 639               597972 2020-11-11  Ontario       251
#> 640               598614 2020-11-12  Ontario       642
#> 641               598784 2020-11-13  Ontario       170
#> 642               598916 2020-11-14  Ontario       132
#> 643               600035 2020-11-15  Ontario      1119
#> 644               600657 2020-11-16  Ontario       622
#> 645               601185 2020-11-17  Ontario       528
#> 646               601775 2020-11-18  Ontario       590
#> 647               602338 2020-11-19  Ontario       563
#> 648               602476 2020-11-20  Ontario       138
#> 649               602687 2020-11-21  Ontario       211
#> 650               603949 2020-11-22  Ontario      1262
#> 651               604662 2020-11-23  Ontario       713
#> 652               605269 2020-11-24  Ontario       607
#> 653               605894 2020-11-25  Ontario       625
#> 654               606539 2020-11-26  Ontario       645
#> 655               606718 2020-11-27  Ontario       179
#> 656               606882 2020-11-28  Ontario       164
#> 657               608238 2020-11-29  Ontario      1356
#> 658               608886 2020-11-30  Ontario       648
#> 659               609731 2020-12-01  Ontario       845
#> 660               610619 2020-12-02  Ontario       888
#> 661               611324 2020-12-03  Ontario       705
#> 662               611556 2020-12-04  Ontario       232
#> 663               611784 2020-12-05  Ontario       228
#> 664               613531 2020-12-06  Ontario      1747
#> 665               614528 2020-12-07  Ontario       997
#> 666               615541 2020-12-08  Ontario      1013
#> 667               616359 2020-12-09  Ontario       818
#> 668               617188 2020-12-10  Ontario       829
#> 669               617413 2020-12-11  Ontario       225
#> 670               617639 2020-12-12  Ontario       226
#> 671               619738 2020-12-13  Ontario      2099
#> 672               620733 2020-12-14  Ontario       995
#> 673               621838 2020-12-15  Ontario      1105
#> 674               622962 2020-12-16  Ontario      1124
#> 675               624412 2020-12-17  Ontario      1450
#> 676               624890 2020-12-18  Ontario       478
#> 677               625230 2020-12-19  Ontario       340
#> 678               627934 2020-12-20  Ontario      2704
#> 679               629766 2020-12-21  Ontario      1832
#> 680               631642 2020-12-22  Ontario      1876
#> 681               633888 2020-12-23  Ontario      2246
#> 682               635967 2020-12-24  Ontario      2079
#> 683               636119 2020-12-25  Ontario       152
#> 684               636226 2020-12-26  Ontario       107
#> 685               640679 2020-12-27  Ontario      4453
#> 686               643105 2020-12-28  Ontario      2426
#> 687               653097 2020-12-29  Ontario      9992
#> 688               658331 2020-12-30  Ontario      5234
#> 689               664236 2020-12-31  Ontario      5905
get_recoveries(loc = "prov", date = "2021-09-01")
#> No encoding supplied: defaulting to UTF-8.
#>    cumulative_recovered       date      province recovered
#> 1                239572 2020-09-01       Alberta       677
#> 2                158900 2020-09-01            BC       959
#> 3                 57074 2020-09-01      Manitoba        36
#> 4                  2600 2020-09-01 New Brunswick        21
#> 5                  1445 2020-09-01            NL         2
#> 6                  5884 2020-09-01   Nova Scotia        12
#> 7                   653 2020-09-01       Nunavut         0
#> 8                   345 2020-09-01           NWT         0
#> 9                550829 2020-09-01       Ontario       650
#> 10                  225 2020-09-01           PEI         3
#> 11               373897 2020-09-01        Quebec       359
#> 12                   13 2020-09-01   Repatriated         0
#> 13                51709 2020-09-01  Saskatchewan       149
#> 14                  662 2020-09-01         Yukon         0