Query total cumulative vaccine completion with ability to specify location and date range of returned data.

Query total cumulative vaccine completion with ability to specify location and date range of returned data.

get_vaccinations(
  loc = "prov",
  date = NULL,
  after = "2020-01-01",
  before = Sys.Date(),
  datetime_type = TRUE
)

Arguments

loc

string: Specify geographic filter and aggregation of returned data. Valid loc arguments are: 'canada', 'prov' and two-letter province codes (e.g. 'ON', 'BC', etc.)

date

string: If not None, return data from the specified date YYYY-MM-DD. Superceeds 'after' and 'before' parameters.

after

string: Return data on and after the specified date YYYY-MM-DD.

before

string: Return data on and before the specified date YYYY-MM-DD.

datetime_type

boolean: Return date column as a character (FALSE) or as a Date (TRUE).

Value

Dataframe containing content of API response.

Examples

get_vaccinations(loc = "ON", before = "2021-12-31")
#> No encoding supplied: defaulting to UTF-8.
#>     cumulative_cvaccine cvaccine       date province
#> 1                  3023     3023 2020-01-12  Ontario
#> 2                  4389     1366 2020-01-13  Ontario
#> 3                  6647     2258 2020-01-14  Ontario
#> 4                  8547     1900 2020-01-15  Ontario
#> 5                  9667     1120 2020-01-16  Ontario
#> 6                 10857     1190 2020-01-17  Ontario
#> 7                 10876       19 2020-01-18  Ontario
#> 8                 12805     1929 2020-01-19  Ontario
#> 9                 16181     3376 2020-01-20  Ontario
#> 10                20113     3932 2020-01-21  Ontario
#> 11                24646     4533 2020-01-22  Ontario
#> 12                28954     4308 2020-01-23  Ontario
#> 13                31441     2487 2020-01-24  Ontario
#> 14                35628     4187 2020-01-25  Ontario
#> 15                41643     6015 2020-01-26  Ontario
#> 16                48230     6587 2020-01-27  Ontario
#> 17                55286     7056 2020-01-28  Ontario
#> 18                61679     6393 2020-01-29  Ontario
#> 19                67787     6108 2020-01-30  Ontario
#> 20                68849     1062 2020-01-31  Ontario
#> 21                70293     1444 2020-02-01  Ontario
#> 22                72057     1764 2020-02-02  Ontario
#> 23                74994     2937 2020-02-03  Ontario
#> 24                80977     5983 2020-02-04  Ontario
#> 25                87831     6854 2020-02-05  Ontario
#> 26                95573     7742 2020-02-06  Ontario
#> 27               101744     6171 2020-02-07  Ontario
#> 28               106163     4419 2020-02-08  Ontario
#> 29               115529     9366 2020-02-09  Ontario
#> 30               125725    10196 2020-02-10  Ontario
#> 31               136988    11263 2020-02-11  Ontario
#> 32               151282    14294 2020-02-12  Ontario
#> 33               164307    13025 2020-02-13  Ontario
#> 34               174643    10336 2020-02-14  Ontario
#> 35               174643        0 2020-02-15  Ontario
#> 36               186934    12291 2020-02-16  Ontario
#> 37               195366     8432 2020-02-17  Ontario
#> 38               205802    10436 2020-02-18  Ontario
#> 39               217715    11913 2020-02-19  Ontario
#> 40               229151    11436 2020-02-20  Ontario
#> 41               235922     6771 2020-02-21  Ontario
#> 42               240669     4747 2020-02-22  Ontario
#> 43               247042     6373 2020-02-23  Ontario
#> 44               251590     4548 2020-02-24  Ontario
#> 45               255449     3859 2020-02-25  Ontario
#> 46               258014     2565 2020-02-26  Ontario
#> 47               260972     2958 2020-02-27  Ontario
#> 48               262103     1131 2020-02-28  Ontario
#> 49               263214     1111 2020-03-01  Ontario
#> 50               264896     1682 2020-03-02  Ontario
#> 51               266710     1814 2020-03-03  Ontario
#> 52               268118     1408 2020-03-04  Ontario
#> 53               269063      945 2020-03-05  Ontario
#> 54               270625     1562 2020-03-06  Ontario
#> 55               271807     1182 2020-03-07  Ontario
#> 56               273676     1869 2020-03-08  Ontario
#> 57               276193     2517 2020-03-09  Ontario
#> 58               279204     3011 2020-03-10  Ontario
#> 59               281714     2510 2020-03-11  Ontario
#> 60               282748     1034 2020-03-12  Ontario
#> 61               284686     1938 2020-03-13  Ontario
#> 62               285615      929 2020-03-14  Ontario
#> 63               287283     1668 2020-03-15  Ontario
#> 64               288918     1635 2020-03-16  Ontario
#> 65               290659     1741 2020-03-17  Ontario
#> 66               292269     1610 2020-03-18  Ontario
#> 67               294749     2480 2020-03-19  Ontario
#> 68               297134     2385 2020-03-20  Ontario
#> 69               298549     1415 2020-03-21  Ontario
#> 70               299297      748 2020-03-22  Ontario
#> 71               301043     1746 2020-03-23  Ontario
#> 72               302664     1621 2020-03-24  Ontario
#> 73               304386     1722 2020-03-25  Ontario
#> 74               306373     1987 2020-03-26  Ontario
#> 75               308301     1928 2020-03-27  Ontario
#> 76               309285      984 2020-03-28  Ontario
#> 77               311248     1963 2020-03-29  Ontario
#> 78               313889     2641 2020-03-30  Ontario
#> 79               315820     1931 2020-03-31  Ontario
#> 80               317715     1895 2020-04-01  Ontario
#> 81               317715        0 2020-04-02  Ontario
#> 82               321469     3754 2020-04-03  Ontario
#> 83               321469        0 2020-04-04  Ontario
#> 84               322197      728 2020-04-05  Ontario
#> 85               323148      951 2020-04-06  Ontario
#> 86               324783     1635 2020-04-07  Ontario
#> 87               326360     1577 2020-04-08  Ontario
#> 88               328598     2238 2020-04-09  Ontario
#> 89               330982     2384 2020-04-10  Ontario
#> 90               333150     2168 2020-04-11  Ontario
#> 91               333419      269 2020-04-12  Ontario
#> 92               335262     1843 2020-04-13  Ontario
#> 93               337206     1944 2020-04-14  Ontario
#> 94               339491     2285 2020-04-15  Ontario
#> 95               341933     2442 2020-04-16  Ontario
#> 96               344244     2311 2020-04-17  Ontario
#> 97               345310     1066 2020-04-18  Ontario
#> 98               346005      695 2020-04-19  Ontario
#> 99               347597     1592 2020-04-20  Ontario
#> 100              349396     1799 2020-04-21  Ontario
#> 101              351354     1958 2020-04-22  Ontario
#> 102              355208     3854 2020-04-23  Ontario
#> 103              358776     3568 2020-04-24  Ontario
#> 104              360354     1578 2020-04-25  Ontario
#> 105              361166      812 2020-04-26  Ontario
#> 106              362563     1397 2020-04-27  Ontario
#> 107              365166     2603 2020-04-28  Ontario
#> 108              368403     3237 2020-04-29  Ontario
#> 109              371110     2707 2020-04-30  Ontario
#> 110              373559     2449 2020-05-01  Ontario
#> 111              375280     1721 2020-05-02  Ontario
#> 112              375905      625 2020-05-03  Ontario
#> 113              378085     2180 2020-05-04  Ontario
#> 114              381123     3038 2020-05-05  Ontario
#> 115              384589     3466 2020-05-06  Ontario
#> 116              387484     2895 2020-05-07  Ontario
#> 117              390990     3506 2020-05-08  Ontario
#> 118              392835     1845 2020-05-09  Ontario
#> 119              393884     1049 2020-05-10  Ontario
#> 120              396787     2903 2020-05-11  Ontario
#> 121              402258     5471 2020-05-12  Ontario
#> 122              407600     5342 2020-05-13  Ontario
#> 123              415531     7931 2020-05-14  Ontario
#> 124              422960     7429 2020-05-15  Ontario
#> 125              429636     6676 2020-05-16  Ontario
#> 126              432760     3124 2020-05-17  Ontario
#> 127              442102     9342 2020-05-18  Ontario
#> 128              456784    14682 2020-05-19  Ontario
#> 129              473759    16975 2020-05-20  Ontario
#> 130              495757    21998 2020-05-21  Ontario
#> 131              519981    24224 2020-05-22  Ontario
#> 132              531603    11622 2020-05-23  Ontario
#> 133              531603        0 2020-05-24  Ontario
#> 134              544288    12685 2020-05-25  Ontario
#> 135              569317    25029 2020-05-26  Ontario
#> 136              594854    25537 2020-05-27  Ontario
#> 137              624920    30066 2020-05-28  Ontario
#> 138              659139    34219 2020-05-29  Ontario
#> 139              687894    28755 2020-05-30  Ontario
#> 140              706832    18938 2020-05-31  Ontario
#> 141              739722    32890 2020-06-01  Ontario
#> 142              781163    41441 2020-06-02  Ontario
#> 143              834981    53818 2020-06-03  Ontario
#> 144              896065    61084 2020-06-04  Ontario
#> 145              963257    67192 2020-06-05  Ontario
#> 146             1031520    68263 2020-06-06  Ontario
#> 147             1086805    55285 2020-06-07  Ontario
#> 148             1174330    87525 2020-06-08  Ontario
#> 149             1282676   108346 2020-06-09  Ontario
#> 150             1399776   117100 2020-06-10  Ontario
#> 151             1533742   133966 2020-06-11  Ontario
#> 152             1667761   134019 2020-06-12  Ontario
#> 153             1796782   129021 2020-06-13  Ontario
#> 154             1894320    97538 2020-06-14  Ontario
#> 155             2037751   143431 2020-06-15  Ontario
#> 156             2198715   160964 2020-06-16  Ontario
#> 157             2371035   172320 2020-06-17  Ontario
#> 158             2547241   176206 2020-06-18  Ontario
#> 159             2725302   178061 2020-06-19  Ontario
#> 160             2874280   148978 2020-06-20  Ontario
#> 161             2972700    98420 2020-06-21  Ontario
#> 162             3145372   172672 2020-06-22  Ontario
#> 163             3343744   198372 2020-06-23  Ontario
#> 164             3542439   198695 2020-06-24  Ontario
#> 165             3761221   218782 2020-06-25  Ontario
#> 166             3988105   226884 2020-06-26  Ontario
#> 167             4163621   175516 2020-06-27  Ontario
#> 168             4324770   161149 2020-06-28  Ontario
#> 169             4563469   238699 2020-06-29  Ontario
#> 170             4808170   244701 2020-06-30  Ontario
#> 171             4808170        0 2020-07-01  Ontario
#> 172             5185939   377769 2020-07-02  Ontario
#> 173             5377506   191567 2020-07-03  Ontario
#> 174             5554637   177131 2020-07-04  Ontario
#> 175             5687477   132840 2020-07-05  Ontario
#> 176             5883641   196164 2020-07-06  Ontario
#> 177             6069647   186006 2020-07-07  Ontario
#> 178             6312906   243259 2020-07-08  Ontario
#> 179             6515960   203054 2020-07-09  Ontario
#> 180             6702624   186664 2020-07-10  Ontario
#> 181             6855946   153322 2020-07-11  Ontario
#> 182             6966273   110327 2020-07-12  Ontario
#> 183             7124007   157734 2020-07-13  Ontario
#> 184             7283528   159521 2020-07-14  Ontario
#> 185             7429926   146398 2020-07-15  Ontario
#> 186             7578116   148190 2020-07-16  Ontario
#> 187             7726829   148713 2020-07-17  Ontario
#> 188             7845691   118862 2020-07-18  Ontario
#> 189             7926705    81014 2020-07-19  Ontario
#> 190             8042242   115537 2020-07-20  Ontario
#> 191             8165247   123005 2020-07-21  Ontario
#> 192             8272485   107238 2020-07-22  Ontario
#> 193             8375133   102648 2020-07-23  Ontario
#> 194             8480761   105628 2020-07-24  Ontario
#> 195             8569752    88991 2020-07-25  Ontario
#> 196             8625932    56180 2020-07-26  Ontario
#> 197             8702111    76179 2020-07-27  Ontario
#> 198             8779459    77348 2020-07-28  Ontario
#> 199             8851757    72298 2020-07-29  Ontario
#> 200             8919362    67605 2020-07-30  Ontario
#> 201             8985638    66276 2020-07-31  Ontario
#> 202             9035335    49697 2020-08-01  Ontario
#> 203             9035335        0 2020-08-02  Ontario
#> 204             9095394    60059 2020-08-03  Ontario
#> 205             9143402    48008 2020-08-04  Ontario
#> 206             9189535    46133 2020-08-05  Ontario
#> 207             9231765    42230 2020-08-06  Ontario
#> 208             9280698    48933 2020-08-07  Ontario
#> 209             9318953    38255 2020-08-08  Ontario
#> 210             9343260    24307 2020-08-09  Ontario
#> 211             9381122    37862 2020-08-10  Ontario
#> 212             9418632    37510 2020-08-11  Ontario
#> 213             9456663    38031 2020-08-12  Ontario
#> 214             9493726    37063 2020-08-13  Ontario
#> 215             9532105    38379 2020-08-14  Ontario
#> 216             9563420    31315 2020-08-15  Ontario
#> 217             9583506    20086 2020-08-16  Ontario
#> 218             9612865    29359 2020-08-17  Ontario
#> 219             9641986    29121 2020-08-18  Ontario
#> 220             9670921    28935 2020-08-19  Ontario
#> 221             9700285    29364 2020-08-20  Ontario
#> 222             9731926    31641 2020-08-21  Ontario
#> 223             9754737    22811 2020-08-22  Ontario
#> 224             9770389    15652 2020-08-23  Ontario
#> 225             9793814    23425 2020-08-24  Ontario
#> 226             9819677    25863 2020-08-25  Ontario
#> 227             9844659    24982 2020-08-26  Ontario
#> 228             9867920    23261 2020-08-27  Ontario
#> 229             9892465    24545 2020-08-28  Ontario
#> 230             9911914    19449 2020-08-29  Ontario
#> 231             9923606    11692 2020-08-30  Ontario
#> 232             9942834    19228 2020-08-31  Ontario
#> 233             9964336    21502 2020-09-01  Ontario
#> 234             9983796    19460 2020-09-02  Ontario
#> 235            10006367    22571 2020-09-03  Ontario
#> 236            10031003    24636 2020-09-04  Ontario
#> 237            10049092    18089 2020-09-05  Ontario
#> 238            10049092        0 2020-09-06  Ontario
#> 239            10068507    19415 2020-09-07  Ontario
#> 240            10089234    20727 2020-09-08  Ontario
#> 241            10109582    20348 2020-09-09  Ontario
#> 242            10128949    19367 2020-09-10  Ontario
#> 243            10152637    23688 2020-09-11  Ontario
#> 244            10170086    17449 2020-09-12  Ontario
#> 245            10179312     9226 2020-09-13  Ontario
#> 246            10195431    16119 2020-09-14  Ontario
#> 247            10215951    20520 2020-09-15  Ontario
#> 248            10236143    20192 2020-09-16  Ontario
#> 249            10256563    20420 2020-09-17  Ontario
#> 250            10280570    24007 2020-09-18  Ontario
#> 251            10299445    18875 2020-09-19  Ontario
#> 252            10309713    10268 2020-09-20  Ontario
#> 253            10327192    17479 2020-09-21  Ontario
#> 254            10351311    24119 2020-09-22  Ontario
#> 255            10378380    27069 2020-09-23  Ontario
#> 256            10406874    28494 2020-09-24  Ontario
#> 257            10437551    30677 2020-09-25  Ontario
#> 258            10461784    24233 2020-09-26  Ontario
#> 259            10474837    13053 2020-09-27  Ontario
#> 260            10493722    18885 2020-09-28  Ontario
#> 261            10517103    23381 2020-09-29  Ontario
#> 262            10517103        0 2020-09-30  Ontario
#> 263            10558963    41860 2020-10-01  Ontario
#> 264            10583567    24604 2020-10-02  Ontario
#> 265            10599457    15890 2020-10-03  Ontario
#> 266            10612380    12923 2020-10-04  Ontario
#> 267            10629132    16752 2020-10-05  Ontario
#> 268            10649353    20221 2020-10-06  Ontario
#> 269            10665079    15726 2020-10-07  Ontario
#> 270            10684356    19277 2020-10-08  Ontario
#> 271            10707708    23352 2020-10-09  Ontario
#> 272            10726674    18966 2020-10-10  Ontario
#> 273            10726674        0 2020-10-11  Ontario
#> 274            10738052    11378 2020-10-12  Ontario
#> 275            10756935    18883 2020-10-13  Ontario
#> 276            10775345    18410 2020-10-14  Ontario
#> 277            10794769    19424 2020-10-15  Ontario
#> 278            10815108    20339 2020-10-16  Ontario
#> 279            10830407    15299 2020-10-17  Ontario
#> 280            10838750     8343 2020-10-18  Ontario
#> 281            10852509    13759 2020-10-19  Ontario
#> 282            10868680    16171 2020-10-20  Ontario
#> 283            10883221    14541 2020-10-21  Ontario
#> 284            10900029    16808 2020-10-22  Ontario
#> 285            10918231    18202 2020-10-23  Ontario
#> 286            10931045    12814 2020-10-24  Ontario
#> 287            10937705     6660 2020-10-25  Ontario
#> 288            10949674    11969 2020-10-26  Ontario
#> 289            10964047    14373 2020-10-27  Ontario
#> 290            10977572    13525 2020-10-28  Ontario
#> 291            10991730    14158 2020-10-29  Ontario
#> 292            11007022    15292 2020-10-30  Ontario
#> 293            11017767    10745 2020-10-31  Ontario
#> 294            11022859     5092 2020-11-01  Ontario
#> 295            11031522     8663 2020-11-02  Ontario
#> 296            11042298    10776 2020-11-03  Ontario
#> 297            11053036    10738 2020-11-04  Ontario
#> 298            11063495    10459 2020-11-05  Ontario
#> 299            11069363     5868 2020-11-06  Ontario
#> 300            11078003     8640 2020-11-07  Ontario
#> 301            11089632    11629 2020-11-08  Ontario
#> 302            11098007     8375 2020-11-09  Ontario
#> 303            11107758     9751 2020-11-10  Ontario
#> 304            11117999    10241 2020-11-11  Ontario
#> 305            11126663     8664 2020-11-12  Ontario
#> 306            11138723    12060 2020-11-13  Ontario
#> 307            11147412     8689 2020-11-14  Ontario
#> 308            11152122     4710 2020-11-15  Ontario
#> 309            11160503     8381 2020-11-16  Ontario
#> 310            11169767     9264 2020-11-17  Ontario
#> 311            11178476     8709 2020-11-18  Ontario
#> 312            11187346     8870 2020-11-19  Ontario
#> 313            11197154     9808 2020-11-20  Ontario
#> 314            11204499     7345 2020-11-21  Ontario
#> 315            11208667     4168 2020-11-22  Ontario
#> 316            11215575     6908 2020-11-23  Ontario
#> 317            11223333     7758 2020-11-24  Ontario
#> 318            11230866     7533 2020-11-25  Ontario
#> 319            11238458     7592 2020-11-26  Ontario
#> 320            11245640     7182 2020-11-27  Ontario
#> 321            11250989     5349 2020-11-28  Ontario
#> 322            11254056     3067 2020-11-29  Ontario
#> 323            11259729     5673 2020-11-30  Ontario
#> 324            11266204     6475 2020-12-01  Ontario
#> 325            11272268     6064 2020-12-02  Ontario
#> 326            11281481     9213 2020-12-03  Ontario
#> 327            11288025     6544 2020-12-04  Ontario
#> 328            11292723     4698 2020-12-05  Ontario
#> 329            11295706     2983 2020-12-06  Ontario
#> 330            11300708     5002 2020-12-07  Ontario
#> 331            11306219     5511 2020-12-08  Ontario
#> 332            11311453     5234 2020-12-09  Ontario
#> 333            11316878     5425 2020-12-10  Ontario
#> 334            11323370     6492 2020-12-11  Ontario
#> 335            11327927     4557 2020-12-12  Ontario
#> 336            11330544     2617 2020-12-13  Ontario
#> 337            11334812     4268 2020-12-14  Ontario
#> 338            11340162     5350 2020-12-15  Ontario
#> 339            11345168     5006 2020-12-16  Ontario
#> 340            11350356     5188 2020-12-17  Ontario
#> 341            11355707     5351 2020-12-18  Ontario
#> 342            11359730     4023 2020-12-19  Ontario
#> 343            11362551     2821 2020-12-20  Ontario
#> 344            11367446     4895 2020-12-21  Ontario
#> 345            11373519     6073 2020-12-22  Ontario
#> 346            11379811     6292 2020-12-23  Ontario
#> 347            11383811     4000 2020-12-24  Ontario
#> 348            11388860     5049 2020-12-25  Ontario
#> 349            11389124      264 2020-12-26  Ontario
#> 350            11390166     1042 2020-12-27  Ontario
#> 351            11394756     4590 2020-12-28  Ontario
#> 352            11400095     5339 2020-12-29  Ontario
#> 353            11404979     4884 2020-12-30  Ontario
#> 354            11410550     5571 2020-12-31  Ontario
get_vaccinations(loc = "prov", date = "2021-09-01")
#> No encoding supplied: defaulting to UTF-8.
#>    cumulative_cvaccine cvaccine       date      province
#> 1              2629789     4240 2020-09-01       Alberta
#> 2              3558964     9209 2020-09-01            BC
#> 3               932193     1372 2020-09-01      Manitoba
#> 4               525603     1853 2020-09-01 New Brunswick
#> 5               367510     3840 2020-09-01            NL
#> 6               690154     1428 2020-09-01   Nova Scotia
#> 7                20887        0 2020-09-01       Nunavut
#> 8                32475        0 2020-09-01           NWT
#> 9              9964336    21502 2020-09-01       Ontario
#> 10              113261        0 2020-09-01           PEI
#> 11             5940781    20875 2020-09-01        Quebec
#> 12              695823     1019 2020-09-01  Saskatchewan
#> 13               31907       45 2020-09-01         Yukon